### 2025-09-23 177425c
Check gate angles in separate method
