+++ src/gates.rs
@@ -100,6 +100,14 @@ pub(crate) fn matrix_adjoint_inplace(data: &mut DataTensor) {
     data.conjugate();
 }
 
+/// Checks the slice has the requested number of elements or panics.
+#[inline]
+fn unpack_angles<const N: usize>(angles: &[f64]) -> [f64; N] {
+    angles
+        .try_into()
+        .unwrap_or_else(|_| panic!("Expected {N} angles, but got {}.", angles.len()))
+}
+
 /// A quantum gate.
 pub trait Gate: Send + Sync {
     /// Returns the name of the gate.
@@ -146,7 +154,7 @@ impl Gate for X {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let z = Complex64::ZERO;
         let o = Complex64::ONE;
         #[rustfmt::skip]
@@ -171,7 +179,7 @@ impl Gate for Y {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let z = Complex64::ZERO;
         let i = Complex64::I;
         #[rustfmt::skip]
@@ -196,7 +204,7 @@ impl Gate for Z {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let z = Complex64::ZERO;
         let o = Complex64::ONE;
         #[rustfmt::skip]
@@ -221,7 +229,7 @@ impl Gate for H {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let h = Complex64::new(FRAC_1_SQRT_2, 0.0);
         #[rustfmt::skip]
         let data = vec![
@@ -245,7 +253,7 @@ impl Gate for T {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let z = Complex64::ZERO;
         let o = Complex64::ONE;
         let t = Complex64::new(FRAC_1_SQRT_2, FRAC_1_SQRT_2);
@@ -273,9 +281,7 @@ impl Gate for U {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        let [theta, phi, lambda] = angles else {
-            panic!("Expected 3 angles, got {}", angles.len())
-        };
+        let [theta, phi, lambda] = unpack_angles(angles);
         let (sin, cos) = (theta / 2.0).sin_cos();
         let data = vec![
             Complex64::new(cos, 0.0),
@@ -288,9 +294,7 @@ impl Gate for U {
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
         // This explicit implementation is ~30% faster
-        let [theta, phi, lambda] = angles else {
-            panic!("Expected 3 angles, got {}", angles.len())
-        };
+        let [theta, phi, lambda] = unpack_angles(angles);
         let (sin, cos) = (theta / 2.0).sin_cos();
         let data = vec![
             Complex64::new(cos, 0.0),
@@ -310,7 +314,7 @@ impl Gate for Sx {
     }
 
     fn compute(&self, angles: &[f64]) -> DataTensor {
-        assert!(angles.is_empty());
+        let [] = unpack_angles(angles);
         let a = Complex64::new(0.5, 0.5);
         let b = Complex64::new(0.5, -0.5);
         #[rustfmt::skip]
[…]
