### 2025-10-01 7d2a9a8
Bump actions-rust-lang/setup-rust-toolchain from 1.14.1 to 1.15.1
Bumps [actions-rust-lang/setup-rust-toolchain](https://github.com/actions-rust-lang/setup-rust-toolchain) from 1.14.1 to 1.15.1.
- [Release notes](https://github.com/actions-rust-lang/setup-rust-toolchain/releases)
- [Changelog](https://github.com/actions-rust-lang/setup-rust-toolchain/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions-rust-lang/setup-rust-toolchain/compare/ac90e63697ac2784f4ecfe2964e1a285c304003a...02be93da58aa71fb456aa9c43b301149248829d8)

---
updated-dependencies:
- dependency-name: actions-rust-l
[…]
