+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
         with:
           submodules: true
       - name: Install stable
-        uses: actions-rust-lang/setup-rust-toolchain@ac90e63697ac2784f4ecfe2964e1a285c304003a # tag=v1.14.1
+        uses: actions-rust-lang/setup-rust-toolchain@02be93da58aa71fb456aa9c43b301149248829d8 # tag=v1.15.1
         with:
           toolchain: stable
       - name: cargo generate-lockfile
@@ -88,11 +88,11 @@ jobs:
         with:
           submodules: true
       - name: Install stable
-        uses: actions-rust-lang/setup-rust-toolchain@ac90e63697ac2784f4ecfe2964e1a285c304003a # tag=v1.14.1
+        uses: actions-rust-lang/setup-rust-toolchain@02be93da58aa71fb456aa9c43b301149248829d8 # tag=v1.15.1
         with:
           toolchain: stable  
       - name: Install nightly for -Zdirect-minimal-versions
-        uses: actions-rust-lang/setup-rust-toolchain@ac90e63697ac2784f4ecfe2964e1a285c304003a # tag=v1.14.1
+        uses: actions-rust-lang/setup-rust-toolchain@02be93da58aa71fb456aa9c43b301149248829d8 # tag=v1.15.1
         with:
           toolchain: nightly
       - name: rustup default stable
@@ -120,7 +120,7 @@ jobs:
   #       with:
   #         submodules: true
   #     - name: Install stable
-  #       uses: actions-rust-lang/setup-rust-toolchain@ac90e63697ac2784f4ecfe2964e1a285c304003a # tag=v1.14.1
+  #       uses: actions-rust-lang/setup-rust-toolchain@02be93da58aa71fb456aa9c43b301149248829d8 # tag=v1.15.1
   #       with:
   #         toolchain: stable
   #     - name: cargo generate-lockfile
@@ -148,7 +148,7 @@ jobs:
         with:
           submodules: true
       - name: Install nightly
-        uses: actions-rust-lang/setup-rust-toolchain@ac90e63697ac2784f4ecfe2964e1a285c304003a # tag=v1.14.1
+        uses: actions-rust-lang/setup-rust-toolchain@02be93da58aa71fb456aa9c43b301149248829d8 # tag=v1.15.1
         with:
           toolchain: nightly
           components: llvm-tools-preview
