+++ paper/paper.md
@@ -12,10 +12,15 @@ authors:
   - name: Qunsheng Huang
     orcid: 0000-0002-1289-6559
     affiliation: '1'
+  - name: Christian B. Mendl
+    orcid: 0000-0002-6386-0230
+    affiliation: '1, 2'
 affiliations:
   - name: School for Computation, Information and Technology, Technical University of Munich, Germany
     index: 1
-date: 8 September 2025
+  - name: Institute for Advanced Study, Technical University of Munich, Germany
+    index: 2
+date: 9 October 2025
 bibliography: paper.bib
 ---
 
@@ -31,27 +36,28 @@ With the rapid advancements in quantum computing, ever larger quantum circuits a
 Since real quantum hardware is often not yet available and has to fight problems such as high error rates, research on efficient classical simulation methods is crucial to bridge this gap.
 Tensor networks have been shown to be a viable tool for classical simulation, disproving even two major claims of quantum advantage, i.e., the realization of an algorithm on real quantum hardware which would be infeasible to simulate on classical hardware [@Pan2022;@Patra2024].
 
-However, existing libraries for tensor network contractions often only consider shared-memory parallelism [@Pfeifer2015], are not open-source [@Bayraktar2023], or are Python-based [@Gray2021] and hence limited in performance.
+However, existing libraries for tensor network contractions often only consider shared-memory parallelism [@Pfeifer2015], are not open-source [@Bayraktar2023], or are written in Python [@Gray2021] and hence less suited for HPC.
 Furthermore, parallelization is usually done by a technique called *slicing*, which incurs computational overhead.
 TNC addresses these limitations by targeting distributed-memory systems, being fully open-source, implemented in Rust, and employing partitioning rather than slicing for parallelization.
 
 # Features
 
-Quantum circuits can be constructed using the library's API or imported from code written in the widely-used OpenQASM2 language [@Cross2017].
-From a given circuit, tensor networks can be constructed that compute the expectation value, the amplitudes to specific measurement outcomes, or the full state vector.
+The user can construct quantum circuits using the library's API or import them from code written in the widely-used OpenQASM2 language [@Cross2017].
+From a given circuit, the user can then generate tensor networks that compute the expectation value, the amplitudes to specific measurement outcomes, or the full state vector.
 Furthermore, random tensor networks can be created as well, e.g., in the form of the original Sycamore experiment [@Arute2019].
 
-For contraction, tensor networks are partitioned based on the number of available MPI ranks.
-The initial partitioning is done using the multi-graph partitioning library KaHyPar [@Andre2018].
-Since the resulting partitionings are often not optimal in terms of the computational cost required to contract them, our library features different algorithms to improve this initial partitioning, such as greedy rebalancing, simulated annealing, and a genetic algorithm.
+For contraction, the library partitions tensor networks based on the number of available MPI ranks.
+It obtains an initial partitioning using the multi-graph partitioning library KaHyPar [@Andre2018].
+Since the resulting partitionings are often not optimal regarding the computational cost required to contract them, our library features different algorithms to improve this initial partitioning, such as greedy rebalancing, simulated annealing, and a genetic algorithm.
 
-Contraction paths, which specify the order of contraction operations, are found using methods from the cotengra library [@Gray2021].
-The individual contractions are then performed using MKL for matrix-matrix multiplications and the hptt library [@Springer2017] for data transposition.
+The library can find contraction paths, which specify the order of contraction operations, using different methods from the cotengra library [@Gray2021].
+It then performs the individual contractions using MKL for matrix-matrix multiplications and the hptt library [@Springer2017] for data transposition.
 
[…]
