### 2025-10-23 #8965 @drvinceknight
@jcmgray @emapuljak @meandmytram I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “TNC: Distributed Tensor Network Contractions in Rust”. You can find more information at the top of this Github issue.
