### 2025-11-01 693fa73
Bump taiki-e/install-action from 2.62.14 to 2.62.43
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.62.14 to 2.62.43.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/3216b6964cbfe053bb8b9a2ef245bd9300e2061d...81ecf985428d5c2ea81dbf079bceca32bc9604ab)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.62.43
  dependency-type: direct:production
  update-type: v
[…]
