+++ .github/workflows/check.yml
@@ -101,7 +101,7 @@ jobs:
         with:
           toolchain: stable
       - name: cargo install cargo-hack
-        uses: taiki-e/install-action@3216b6964cbfe053bb8b9a2ef245bd9300e2061d # tag=v2.62.14
+        uses: taiki-e/install-action@81ecf985428d5c2ea81dbf079bceca32bc9604ab # tag=v2.62.43
         with:
           tool: cargo-hack
       # intentionally no target specifier; see https://github.com/jonhoo/rust-ci-conf/pull/4
