+++ .github/workflows/test.yml
@@ -153,7 +153,7 @@ jobs:
           toolchain: nightly
           components: llvm-tools-preview
       - name: cargo install cargo-llvm-cov
-        uses: taiki-e/install-action@3216b6964cbfe053bb8b9a2ef245bd9300e2061d # tag=v2.62.14
+        uses: taiki-e/install-action@81ecf985428d5c2ea81dbf079bceca32bc9604ab # tag=v2.62.43
         with:
           tool: cargo-llvm-cov
       - name: cargo generate-lockfile
