+++ Cargo.toml
@@ -36,7 +36,7 @@ regex = "1.11.1"
 ndarray = "0.16.1"
 genetic_algorithm = { git = "https://github.com/Ectras/genetic-algorithm.git", branch = "f64_score" }
 rayon = "1.10.0"
-ordered-float = "5.0.0"
+ordered-float = "5.1.0"
 rustengra = { git = "https://github.com/Ectras/rustengra.git", optional = true }
 cotengrust = { git = "https://github.com/Ectras/cotengrust.git", branch = "rust_extension" }
 serde-pickle = "1.2.0"
