### 2025-11-17 9964ce8
Fix rand v0.9 incompatibilities
