### 2025-11-26 #8965 @Luthaf
I don't know much about Tensor Network Contractions, but I can do a review of the Rust code/tests/CI/... !
