### 2025-12-01 bff9355
Bump taiki-e/install-action from 2.62.43 to 2.62.60
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.62.43 to 2.62.60.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/81ecf985428d5c2ea81dbf079bceca32bc9604ab...3575e532701a5fc614b0c842e4119af4cc5fd16d)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.62.60
  dependency-type: direct:production
  update-type: v
[…]
