+++ .github/workflows/test.yml
@@ -153,7 +153,7 @@ jobs:
           toolchain: nightly
           components: llvm-tools-preview
       - name: cargo install cargo-llvm-cov
-        uses: taiki-e/install-action@81ecf985428d5c2ea81dbf079bceca32bc9604ab # tag=v2.62.43
+        uses: taiki-e/install-action@3575e532701a5fc614b0c842e4119af4cc5fd16d # tag=v2.62.60
         with:
           tool: cargo-llvm-cov
       - name: cargo generate-lockfile
