### 2025-12-09 #9598 @editorialbot
[REVIEW]: TNC: Distributed Tensor Network Contractions in Rust
**Submitting author:** @Ectras (<a href="http://orcid.org/0000-0003-3514-8657">Manuel Geiger</a>)
**Repository:** https://github.com/qc-tum/TNC
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
