+++ .github/workflows/test.yml
@@ -31,7 +31,7 @@ jobs:
           python-version: '3.12'
       - name: Install python dependencies
         run: pip install cotengra kahypar optuna
-      - uses: actions/checkout@1af3b93b6815bc44a9784bd300feb67ff0d1eeb3 # tag=v6.0.0
+      - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # tag=v6.0.1
         with:
           submodules: true
       - name: Install stable
@@ -84,7 +84,7 @@ jobs:
           python-version: '3.12'
       - name: Install python dependencies
         run: pip install cotengra kahypar optuna
-      - uses: actions/checkout@1af3b93b6815bc44a9784bd300feb67ff0d1eeb3 # tag=v6.0.0
+      - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # tag=v6.0.1
         with:
           submodules: true
       - name: Install stable
@@ -116,7 +116,7 @@ jobs:
   #     #   if: runner.os == 'Windows'
   #     # - run: vcpkg install openssl:x64-windows-static-md
   #     #   if: runner.os == 'Windows'
-  #     - uses: actions/checkout@1af3b93b6815bc44a9784bd300feb67ff0d1eeb3 # tag=v6.0.0
+  #     - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # tag=v6.0.1
   #       with:
   #         submodules: true
   #     - name: Install stable
@@ -144,7 +144,7 @@ jobs:
           python-version: '3.12'
       - name: Install python dependencies
         run: pip install cotengra kahypar optuna
-      - uses: actions/checkout@1af3b93b6815bc44a9784bd300feb67ff0d1eeb3 # tag=v6.0.0
+      - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # tag=v6.0.1
         with:
           submodules: true
       - name: Install nightly
