### 2026-01-01 e6ce314
Bump taiki-e/install-action from 2.62.60 to 2.65.10
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.62.60 to 2.65.10.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/3575e532701a5fc614b0c842e4119af4cc5fd16d...e0db384ad69f5ba2c6dd0129d8934e0d0ba465c1)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.65.10
  dependency-type: direct:production
  update-type: v
[…]
