+++ .github/workflows/test.yml
@@ -153,7 +153,7 @@ jobs:
           toolchain: nightly
           components: llvm-tools-preview
       - name: cargo install cargo-llvm-cov
-        uses: taiki-e/install-action@3575e532701a5fc614b0c842e4119af4cc5fd16d # tag=v2.62.60
+        uses: taiki-e/install-action@e0db384ad69f5ba2c6dd0129d8934e0d0ba465c1 # tag=v2.65.10
         with:
           tool: cargo-llvm-cov
       - name: cargo generate-lockfile
