### 2026-01-03 #9598 @emapuljak
- [x] **Repository:** Is the source code for this software available at the [https://github.com/qc-tum/TNC](https://github.com/qc-tum/TNC)? **Yes**
- [x] **License:** Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? **dual-licensed under Apache-2.0 and MIT, with a license choice for users**
- [x] **Contribution and authorship:** Has the submitting author (@Ectras) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? - Both @Ectras and @keefehuang have made significant contributions. The list of authors seems complete and appropriate

In the documentation, there is no designated tab "Installation" or "Getting started". I see it just in README. Since its made for HPC linux environments, the installation is not fully supported for macOS, but that's okay, I would say. I would just add `cargo build` to the installation guide.

- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. - Yes, there is a list of dependencies in `.toml` file. But as written above in Functionality -> Installation, a more detailed instalation instructions would be useful, also to be added in documentation.
 - [x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). - (1st) It would be useful to write in README and documentation how to run examples, i.e. - `cargo run --example <example_name>`. (2nd) I tried to run `basic_usage` example, but I'm getting some macOS related errors. I would suggest @Luthaf  to run the examples in Linux if possible.
- [x] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? Documentation is somehow confusing. I would add a high-level overview on the landing page of docs, explaining maybe even visually modules architectural diagram, and also connect tensor networks, contraction paths, partitioning, MPI. Other suggestions: 1) Reorganize sidebar into groups, 2) Add a quickstart (tutorial) guide, 3) enhance visibility of examples (not visible now), 4) maybe it would be useful to explain MPI usage in docs, just in case someone who is not familiar can get a small intro.
- [x] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified? - I would recommend adding more tests: edge cases, determinism when running `opt.optimize_path()` on same TN twice, different failure/error handling modes, extra MPI tests, etc.
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support - **No**, They should add this in README and documentation.

- [x] **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? **The explanation to diverse, non-specialist audience could be improved.**
- [x] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)? I see references in `paper.md` with `@` inside text, and in `paper.bib`, but they are missing as list at the end of the text in `paper.md`.
