### 2026-01-07 e68271b
Integrate types.rs into fitting modules
