### 2026-01-08 b19e2e0
Replace bincode with postcard
bincode is unmaintained, which is a security issue
(https://rustsec.org/advisories/RUSTSEC-2025-0141.html). Hence, this
commit uses postcard, which has comparable features, binary sizes and
speed. bincode is still kept as dev-dependency for the moment, as it was
used to store benchmarking circuits on the disk and we still want to be
able to load them.
