### 2026-01-08 3be058f
Convert into workspace with tnc as member
