+++ tnc/src/qasm/.antlr/Qasm2Lexer.java
@@ -1,0 +1,353 @@
+// Generated from /work_fast/ga87com/QuantumResearch/TensorStuff/GITHUB/tensornetworkcontractions/src/qasm/Qasm2Lexer.g4 by ANTLR 4.13.1
+import org.antlr.v4.runtime.Lexer;
+import org.antlr.v4.runtime.CharStream;
+import org.antlr.v4.runtime.Token;
+import org.antlr.v4.runtime.TokenStream;
+import org.antlr.v4.runtime.*;
+import org.antlr.v4.runtime.atn.*;
+import org.antlr.v4.runtime.dfa.DFA;
+import org.antlr.v4.runtime.misc.*;
+
+@SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast", "CheckReturnValue", "this-escape"})
+public class Qasm2Lexer extends Lexer {
+	static { RuntimeMetaData.checkVersion("4.13.1", RuntimeMetaData.VERSION); }
+
+	protected static final DFA[] _decisionToDFA;
+	protected static final PredictionContextCache _sharedContextCache =
+		new PredictionContextCache();
+	public static final int
+		OPENQASM=1, INCLUDE=2, QREG=3, CREG=4, GATE=5, OPAQUE=6, RESET=7, MEASURE=8, 
+		BARRIER=9, IF=10, PI=11, U=12, CX=13, LBRACKET=14, RBRACKET=15, LBRACE=16, 
+		RBRACE=17, LPAREN=18, RPAREN=19, SEMICOLON=20, COMMA=21, DOT=22, ARROW=23, 
+		EQUALS=24, PLUS=25, MINUS=26, ASTERISK=27, SLASH=28, CARET=29, SIN=30, 
+		COS=31, TAN=32, EXP=33, LN=34, SQRT=35, Integer=36, Float=37, StringLiteral=38, 
+		Whitespace=39, Newline=40, LineComment=41, BlockComment=42, Identifier=43, 
+		VERSION_IDENTIFER_WHITESPACE=44, VersionSpecifier=45;
+	public static final int
+		VERSION_IDENTIFIER=1;
+	public static String[] channelNames = {
+		"DEFAULT_TOKEN_CHANNEL", "HIDDEN"
+	};
+
+	public static String[] modeNames = {
+		"DEFAULT_MODE", "VERSION_IDENTIFIER"
+	};
+
+	private static String[] makeRuleNames() {
+		return new String[] {
+			"OPENQASM", "INCLUDE", "QREG", "CREG", "GATE", "OPAQUE", "RESET", "MEASURE", 
+			"BARRIER", "IF", "PI", "U", "CX", "LBRACKET", "RBRACKET", "LBRACE", "RBRACE", 
+			"LPAREN", "RPAREN", "SEMICOLON", "COMMA", "DOT", "ARROW", "EQUALS", "PLUS", 
+			"MINUS", "ASTERISK", "SLASH", "CARET", "SIN", "COS", "TAN", "EXP", "LN", 
+			"SQRT", "Integer", "FloatLiteralExponent", "Float", "StringLiteral", 
+			"Whitespace", "Newline", "LineComment", "BlockComment", "Identifier", 
[…]
