+++ Cargo.toml
@@ -1,58 +1,3 @@
-[package]
-name = "tnc"
-version = "1.0.0"
-edition = "2021"
-description = "A distributed-memory tensor network contraction library using partitioning."
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/Ectras/TNC"
-keywords = ["tensor networks"]
-categories = ["science"]
-rust-version = "1.85"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-cotengra = ["dep:rustengra"]
-
-[dependencies]
-mpi = { version = "0.8.0", features = ["complex", "derive"] }
-hdf5-metno = { version = "0.11.0", features = ["complex"] }
-itertools = "0.14.0"
-num-complex = "0.4.6"
-rand = "0.9.2"
-antlr-rust = "0.3.0-beta"
-float-cmp = "0.10.0"
-tetra = { git = "https://github.com/Ectras/tetra.git", features = ["serde"] }
-kahypar = { git = "https://github.com/Ectras/rust-kahypar" }
-permutation = "0.4.1"
-serde = { version = "1.0.218", features = ["derive"] }
-postcard = { version = "1.1.3", default-features = false, features = ["use-std"] }
-lazy_static = "1.5.0"
-log = { version = "0.4.22", features = ["kv_std", "kv_serde"] }
-rustc-hash = "2.1.1"
-chrono = "0.4.38"
-serde_json = "1.0.133"
-regex = "1.11.1"
-ndarray = "0.17.1"
-genetic_algorithm = { git = "https://github.com/Ectras/genetic-algorithm.git", branch = "f64_score" }
-rayon = "1.10.0"
-ordered-float = "5.1.0"
-rustengra = { git = "https://github.com/Ectras/rustengra.git", optional = true }
-cotengrust = { git = "https://github.com/Ectras/cotengrust.git", branch = "rust_extension" }
-serde-pickle = "1.2.0"
-jemallocator = "0.5.4"
-
-[dev-dependencies]
-bincode = { version = "2.0.1", features = ["serde"] }
-mpi-test = { git = "https://github.com/Ectras/mpi-test.git" }
-flexi_logger = { version = "0.31.7", features = ["kv", "json", "async"] }
-clap = { version = "4.5.28", features = ["derive"] }
-flate2 = "1.1.1"
-
-[[example]]
-name = "supermuc_sweep"
-required-features = ["cotengra"]
-
-[profile.release]
-codegen-units = 1
-lto = true
+[workspace]
+resolver = "2"
+members = ["tnc"]
