### 2026-01-08 7ea90d2
Move benchmark example into own crate
The benchmark code was not really an example, but an actual use case of
the library. Also, we can remove a few dev-dependencies from the library
and, more importantly, the custom allocator can be pulled out, because
it's probably not so nice to have a library (and not a binary) set the
global allocator.
