+++ benchmark/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "benchmark"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+rand = "0.9.2"
+clap = { version = "4.5.28", features = ["derive"] }
+flate2 = "1.1.1"
+itertools = "0.14.0"
+tnc = { version = "1.0.0", path = "../tnc", features = ["cotengra"] }
+bincode = { version = "2.0.1", features = ["serde"] }
+mpi = { version = "0.8.0", features = ["complex", "derive"] }
+serde_json = "1.0.133"
+serde = { version = "1.0.218", features = ["derive"] }
+flexi_logger = { version = "0.31.7", features = ["kv", "json", "async"] }
+log = { version = "0.4.22", features = ["kv_std", "kv_serde"] }
