### 2026-01-08 c2f923d
Move Jemallocator to benchmark crate
A library should probably not set the global allocator, hence move it to
the binary.
