+++ tnc/src/lib.rs
@@ -3,11 +3,6 @@
 //! The partitionings can then be contracted in parallel on a distributed-memory system, as common in high-performance computing.
 //! Local contractions on a single system are also possible.
 
-extern crate jemallocator;
-
-#[global_allocator]
-static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
-
 pub mod builders;
 pub mod contractionpath;
 pub mod gates;
