+++ tnc/Cargo.toml
@@ -40,7 +40,6 @@ ordered-float = "5.1.0"
 rustengra = { git = "https://github.com/Ectras/rustengra.git", optional = true }
 cotengrust = { git = "https://github.com/Ectras/cotengrust.git", branch = "rust_extension" }
 serde-pickle = "1.2.0"
-jemallocator = "0.5.4"
 
 [dev-dependencies]
 mpi-test = { git = "https://github.com/Ectras/mpi-test.git" }
