+++ benchmark/Cargo.toml
@@ -15,3 +15,4 @@ serde_json = "1.0.133"
 serde = { version = "1.0.218", features = ["derive"] }
 flexi_logger = { version = "0.31.7", features = ["kv", "json", "async"] }
 log = { version = "0.4.22", features = ["kv_std", "kv_serde"] }
+jemallocator = "0.5.4"
