### 2026-01-08 4d2b75d
Fix unused benchmark release profile
