+++ benchmark/Cargo.toml
@@ -10,7 +10,7 @@ flate2 = "1.1.1"
 itertools = "0.14.0"
 tnc = { version = "1.0.0", path = "../tnc", features = ["cotengra"] }
 bincode = { version = "2.0.1", features = ["serde"] }
-mpi = { version = "0.8.0", features = ["complex", "derive"] }
+mpi = "0.8.0"
 serde_json = "1.0.133"
 serde = { version = "1.0.218", features = ["derive"] }
 flexi_logger = { version = "0.31.7", features = ["kv", "json", "async"] }
