+++ benchmark/.cargo/config.toml
@@ -1,0 +1,3 @@
+[profile.release]
+codegen-units = 1
+lto = true
