### 2026-01-09 bf09f97
Merge if statements in benchmark
By extracting the code into a separate crate, this new crate has edition
2024 which introduced the feature of if let statements with normal
boolean conditions, hence allowing this change.
