### 2026-01-09 569213d
Fix doc-rs CI
cargo-docs-rs doesn't support workspaces. Hence make it explicitly check
the docs of the tnc crate (the benchmark could be checked as well, but
I don't think it's necessary).
