### 2026-01-09 1ba4fed
Revert "Merge if statements in benchmark"
This reverts commit bf09f9731394684edf173eeeb531c20625a9a701.
