### 2026-01-09 2f3deb4
Lower benchmark Rust edition to 2021
Using edition 2024 requires at least Rust 1.88.0. Since we don't want to
complicate the CI by running different Rust versions for the different
workspace members, the only other options are raising the MRSV to 1.88.0
or downgrading the benchmark grade to edition 2021. The latter seemed
better.
