+++ benchmark/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "benchmark"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 
 [dependencies]
 rand = "0.9.2"
