+++ benchmark/src/main.rs
@@ -11,10 +11,10 @@ use std::time::{Duration, Instant};
 
 use clap::Parser;
 use cli::{Cli, Mode};
-use flate2::Compression;
 use flate2::read::ZlibDecoder;
 use flate2::write::ZlibEncoder;
-use itertools::{Itertools, iproduct};
+use flate2::Compression;
+use itertools::{iproduct, Itertools};
 use log::info;
 use mpi::topology::SimpleCommunicator;
 use mpi::traits::{Communicator, CommunicatorCollectives};
@@ -23,13 +23,12 @@ use rand::distr::StandardUniform;
 use rand::rngs::StdRng;
 use rand::{Rng, RngCore, SeedableRng};
 use results::{OptimizationResult, RunResult, Writer};
-use tnc::contractionpath::ContractionIndex;
 use tnc::contractionpath::communication_schemes::CommunicationScheme;
 use tnc::contractionpath::contraction_cost::{
     communication_path_cost, compute_memory_requirements, contract_size_tensors_exact,
 };
 use tnc::contractionpath::contraction_tree::balancing::{
-    BalanceSettings, BalancingScheme, balance_partitions_iter,
+    balance_partitions_iter, BalanceSettings, BalancingScheme,
 };
 use tnc::contractionpath::paths::cotengrust::{Cotengrust, OptMethod};
 use tnc::contractionpath::paths::hyperoptimization::{HyperOptions, Hyperoptimizer};
@@ -41,6 +40,7 @@ use tnc::contractionpath::repartitioning::simulated_annealing::{
     NaivePartitioningModel,
 };
 use tnc::contractionpath::repartitioning::{compute_solution, simulated_annealing};
+use tnc::contractionpath::ContractionIndex;
 use tnc::mpi::communication::{
     broadcast_path, broadcast_serializing, extract_communication_path,
     intermediate_reduce_tensor_network, scatter_tensor_network,
