+++ README.md
@@ -152,7 +152,3 @@ fn distributed_contraction(tensor: Tensor, world: &SimpleCommunicator) -> Tensor
 
 ## Publications
 - <i>Optimizing Tensor Network Partitioning using Simulated Annealing</i>, Geiger et al. (2025): <https://arxiv.org/abs/2507.20667>
-
-## Contribution
-
-Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed under the terms of both the Apache License, Version 2.0 and the MIT license without any additional terms or conditions.
