### 2026-01-09 6c99d11
Remove unused Default impl of MessageBinaryBlob
