### 2026-01-09 2337a53
Add simple local contraction example
