### 2026-01-09 41212c0
Rename permutor.is_empty() to .is_identity()
