### 2026-01-09 daa40a8
Rename QASM tensor network methods to circuit methods
