+++ tnc/src/qasm.rs
@@ -1,14 +1,14 @@
 //! Import utilities for OpenQASM 2 code.
 
 mod ast;
+mod circuit_creator;
 mod expression_folder;
 mod expression_simplification;
 mod gate_inliner;
 mod generated;
 mod include_resolver;
 mod parser;
-mod qasm_to_tensornetwork;
-mod tn_creator;
+mod qasm_importer;
 mod utils;
 
-pub use qasm_to_tensornetwork::create_tensornetwork;
+pub use qasm_importer::import_qasm;
