### 2026-01-09 a5c3461
Add some documentation to benchmark crate
