### 2026-01-12 1c28b30
Add tutorial on pathfinding
