+++ tnc/src/_tutorial.rs
@@ -1,1 +1,2 @@
 pub mod constructing_tns;
+pub mod pathfinding_and_contraction;
