### 2026-01-13 6907e0c
Add tutorial on tensor structure
