+++ tnc/src/_tutorial.rs
@@ -1,2 +1,3 @@
 pub mod constructing_tns;
 pub mod pathfinding_and_contraction;
+pub mod tensors_structure;
