+++ tnc/src/_tutorial.rs
@@ -1,3 +1,4 @@
 pub mod constructing_tns;
+pub mod hpc_overview;
 pub mod pathfinding_and_contraction;
 pub mod tensors_structure;
