### 2026-01-16 6721173
Rename optimize_path to find_path
The name `optimize_path` was misleading as it suggested that an existing
path is being optimized. However, the method searches for a new path, so
`find_path` seems like a better fit.
