### 2026-01-16 ad78c43
Add ContractionPath struct
This makes handling hierarchical contraction paths a bit nicer. In
particular, it clearly separates the nested paths and the top-level
path, rendering code to e.g. extract the top-level path unnecessary.
At the same type, it gives a bit more type expressiveness, as it can now
be expressed whether a flat path (SimplePath) is taken / returned or a
hierarchical path (ContractionPath).
