+++ tnc/src/contractionpath/paths/hyperoptimization.rs
@@ -4,7 +4,6 @@ use std::{
     time::Duration,
 };
 
-use itertools::Itertools;
 use rustc_hash::FxHashMap;
 use serde::Serialize;
 use serde_pickle::{DeOptions, SerOptions};
@@ -13,7 +12,7 @@ use crate::{
     contractionpath::{
         contraction_cost::contract_path_cost,
         paths::{CostType, FindPath},
-        ssa_replace_ordering, ContractionIndex,
+        ssa_replace_ordering, ContractionPath,
     },
     tensornetwork::tensor::Tensor,
 };
@@ -25,7 +24,7 @@ pub struct Hyperoptimizer<'a> {
     hyper_options: HyperOptions,
     best_flops: f64,
     best_size: f64,
-    best_path: Vec<ContractionIndex>,
+    best_path: ContractionPath,
 }
 
 impl<'a> Hyperoptimizer<'a> {
@@ -40,7 +39,7 @@ impl<'a> Hyperoptimizer<'a> {
             hyper_options,
             best_flops: f64::INFINITY,
             best_size: f64::INFINITY,
-            best_path: vec![],
+            best_path: ContractionPath::default(),
         }
     }
 }
@@ -146,10 +145,7 @@ impl FindPath for Hyperoptimizer<'_> {
 
         let ssa_path = python_hyperoptimizer(&inputs, &outputs, &size_dict, &self.hyper_options);
 
-        self.best_path = ssa_path
-            .iter()
-            .map(|(i, j)| ContractionIndex::Pair(*i, *j))
-            .collect_vec();
+        self.best_path = ContractionPath::simple(ssa_path);
 
         let (op_cost, mem_cost) =
             contract_path_cost(self.tensor.tensors(), &self.get_best_replace_path(), true);
@@ -166,12 +162,12 @@ impl FindPath for Hyperoptimizer<'_> {
         self.best_size
     }
 
-    fn get_best_path(&self) -> &Vec<ContractionIndex> {
+    fn get_best_path(&self) -> &ContractionPath {
         &self.best_path
     }
 
-    fn get_best_replace_path(&self) -> Vec<ContractionIndex> {
-        ssa_replace_ordering(&self.best_path, self.tensor.tensors().len())
+    fn get_best_replace_path(&self) -> ContractionPath {
+        ssa_replace_ordering(&self.best_path)
     }
 }
 
