### 2026-01-20 f7b23d3
Move HDF5 file documentation into module
