+++ tnc/examples/basic_usage.rs
@@ -1,111 +1,102 @@
-use flexi_logger::{json_format, Duplicate, FileSpec, Logger};
-use log::{debug, info, LevelFilter};
-use mpi::traits::Communicator;
-use mpi::Rank;
-use rand::rngs::StdRng;
-use rand::SeedableRng;
-use tnc::builders::connectivity::ConnectivityLayout;
-use tnc::builders::random_circuit::random_circuit;
-use tnc::contractionpath::contraction_cost::contract_cost_tensors;
-use tnc::contractionpath::contraction_tree::export::{to_dendogram_format, to_pdf};
-use tnc::contractionpath::contraction_tree::ContractionTree;
-use tnc::contractionpath::paths::cotengrust::{Cotengrust, OptMethod};
-use tnc::contractionpath::paths::FindPath;
-use tnc::mpi::communication::{
-    broadcast_path, intermediate_reduce_tensor_network, scatter_tensor_network,
+use std::fs;
+
+use mpi::{topology::SimpleCommunicator, traits::Communicator};
+use tnc::{
+    contractionpath::paths::{
+        cotengrust::{Cotengrust, OptMethod},
+        FindPath,
+    },
+    mpi::communication::{
+        broadcast_path, intermediate_reduce_tensor_network, scatter_tensor_network,
+    },
+    qasm::import_qasm,
+    tensornetwork::{
+        contraction::contract_tensor_network,
+        partitioning::{
+            find_partitioning, partition_config::PartitioningStrategy, partition_tensor_network,
+        },
+        tensor::Tensor,
+    },
 };
-use tnc::tensornetwork::contraction::contract_tensor_network;
-use tnc::tensornetwork::partitioning::partition_config::PartitioningStrategy;
-use tnc::tensornetwork::partitioning::{find_partitioning, partition_tensor_network};
 
-/// Sets up logging for rank `rank`. Each rank logs to a separate file and to stdout.
-fn setup_logging_mpi(rank: Rank) {
-    let _logger = Logger::with(LevelFilter::Debug)
-        .format(json_format)
-        .log_to_file(
-            FileSpec::default()
-                .discriminant(format!("rank{rank}"))
-                .suppress_timestamp()
-                .suffix("log.json"),
-        )
-        .duplicate_to_stdout(Duplicate::Info)
-        .start()
-        .unwrap();
-}
-
-// Run with at least 2 processes
 fn main() {
-    let mut rng = StdRng::seed_from_u64(23);
+    // Read from file
+    let tensor = read_qasm("foo.qasm");
+
+    // Set up MPI
     let universe = mpi::initialize().unwrap();
     let world = universe.world();
-    let size = world.size();
     let rank = world.rank();
+    let size = world.size();
+
+    // Perform the contraction
+    let result = if size == 1 {
+        local_contraction(tensor)
+    } else {
+        distributed_contraction(tensor, &world)
+    };
+
+    // Print the result
+    if rank == 0 {
+        println!("{result:?}");
+    }
+}
+
+fn read_qasm(file: &str) -> Tensor {
+    let source = fs::read_to_string(file).unwrap();
+    let circuit = import_qasm(source);
+    let tensor = circuit.into_expectation_value_network();
+    tensor
+}
+
+fn local_contraction(tensor: Tensor) -> Tensor {
+    // Find a contraction path for the whole network
+    let mut opt = Cotengrust::new(&tensor, OptMethod::Greedy);
+    opt.find_path();
+    let contract_path = opt.get_best_replace_path();
+
+    // Contract the whole tensor network on this single node
+    contract_tensor_network(tensor, &contract_path)
+}
+
[…]
