+++ tnc/src/_tutorial.rs
@@ -1,3 +1,12 @@
+//! # Tutorial
+//!
+//! Here you will find explanations and instructions on various topics related to the
+//! library. If you are missing something, feel free to open an issue.
+//!
+//! Please note that we only include code examples here when necessary, as doc tests
+//! are unfortunately only executed for code examples in function docstrings (and not
+//! module docstrings). Accordingly, code examples in the tutorials are not executed
+//! and therefore run the risk of quickly becoming outdated.
 pub mod constructing_tns;
 pub mod hpc_overview;
 pub mod pathfinding_and_contraction;
