+++ tnc/src/mpi.rs
@@ -1,4 +1,5 @@
-//! Methods for MPI interaction.
+//! Methods for MPI interaction, i.e., communication and synchronization between
+//! multiple compute nodes.
 
 pub mod communication;
 mod mpi_types;
