+++ tnc/src/tensornetwork.rs
@@ -1,3 +1,4 @@
+//! Code concerning the tensor datastructure itself.
 pub mod contraction;
 pub mod partitioning;
 pub mod tensor;
