### 2026-01-20 12e088a
Make TensorContraction trait private
It is only used in the contraction module and I don't see any uses cases
outside of this.
