### 2026-01-20 38e487d
Apply clippy lints to basic_usage example
