### 2026-01-20 91bebb5
Apply pedantic clippy lints
