+++ tnc/src/tensornetwork/partitioning.rs
@@ -94,7 +94,6 @@ pub fn find_partitioning(
 /// # Arguments
 ///
 /// * `tensors` - &[(usize, `Tensor`)] to be partitioned. each tuple contains the intermediate contraction cost and intermediate tensor for communication.
-/// * `bond_dims` - bond_dims for tensors
 /// * `k` - number of partitions
 /// * `partitioning_strategy` - The strategy to pass to `KaHyPar`
 /// * `min` - if `true` performs `min_cut` to partition tensor network, if `false`, uses `max_cut`
