### 2026-01-20 d48fd6b
Replace lazy_static with std::sync::LazyLock
