+++ tnc/Cargo.toml
@@ -27,7 +27,6 @@ kahypar = { git = "https://github.com/Ectras/rust-kahypar" }
 permutation = "0.4.1"
 serde = { version = "1.0.218", features = ["derive"] }
 postcard = { version = "1.1.3", default-features = false, features = ["use-std"] }
-lazy_static = "1.5.0"
 log = { version = "0.4.22", features = ["kv_std", "kv_serde"] }
 rustc-hash = "2.1.1"
 chrono = "0.4.38"
