### 2026-01-21 #9598 @danielskatz
Hi - As the track editor, I wanted to check on how this is going, as I don't see that anything has changed in the last few weeks. @drvinceknight - Can you update me on the overall status here and the next steps?
