+++ tnc/Cargo.toml
@@ -15,7 +15,7 @@ rust-version = "1.85"
 cotengra = ["dep:rustengra"]
 
 [dependencies]
-mpi = { version = "0.8.0", features = ["complex", "derive"] }
+mpi = { version = "0.8.1", features = ["complex", "derive"] }
 hdf5-metno = { version = "0.11.0", features = ["complex"] }
 itertools = "0.14.0"
 num-complex = "0.4.6"
