### 2026-01-23 f985bf5
Remove dendogram import/export
The code was not of high quality and not tested. While it was useful for
research, it is also nothing that should be part of the library. A
better place would be a separate crate.
