+++ tnc/src/contractionpath/contraction_tree/import.rs
@@ -1,436 +1,0 @@
-//! Functionality to import the logs of an MPI run to produce a dendogram of the
-//! actual contraction timings.
-//!
-//! Requires to run the code with log level DEBUG. Writing the logs asynchronously
-//! should hopefully reduce the performance impact of the extended logging.
-//!
-//! This code was written quick-and-dirty and might be removed or reworked in the
-//! future.
-
-use std::{fs, rc::Rc};
-
-use chrono::{DateTime, Utc};
-use itertools::Itertools;
-use regex::RegexSet;
-use rustc_hash::FxHashMap;
-
-use crate::contractionpath::contraction_tree::{
-    export::{to_pdf, DendogramEntry, COLORS, COMMUNICATION_COLOR},
-    node::{child_node, parent_node, NodeRef},
-    ContractionTree,
-};
-use crate::utils::traits::HashMapInsertNew;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Direction {
-    Send,
-    Recv,
-}
-
-pub fn logs_to_pdf(filename: &str, suffix: &str, ranks: usize, output: &str) {
-    let height = 120.;
-    let width = 80.;
-    let LogsToTreeResult {
-        tree: contraction_tree,
-        tensor_cost: tensor_position,
-        tensor_color,
-        communication_data: mut communication_logging,
-    } = logs_to_tree(filename, suffix, ranks);
-
-    let flat_contraction_path =
-        contraction_tree.to_flat_contraction_path(contraction_tree.root_id().unwrap(), false);
-
-    let leaf_nodes = contraction_tree.leaf_ids(contraction_tree.root_id().unwrap());
-
-    let num_leaf_nodes = leaf_nodes.len();
-    let mut dendogram_entries = Vec::new();
-
-    let mut next_x = 0f64;
-    let mut tensor_x_position = FxHashMap::default();
-
-    for (i, j) in &flat_contraction_path {
-        if contraction_tree.node(*i).is_leaf() {
-            tensor_x_position.insert_new(*i, next_x);
-            dendogram_entries.push(DendogramEntry {
-                id: *i,
-                x: tensor_x_position[i],
-                y: 0f64,
-                cost: 0f64,
-                color: tensor_color[i].clone(),
-                children: None,
-            });
-            next_x += 1f64;
-        }
-
-        if contraction_tree.node(*j).is_leaf() {
-            tensor_x_position.insert_new(*j, next_x);
-            dendogram_entries.push(DendogramEntry {
-                id: *j,
-                x: tensor_x_position[j],
-                y: 0f64,
-                cost: 0f64,
-                color: tensor_color[j].clone(),
-                children: None,
-            });
-            next_x += 1f64;
-        }
-
-        let x1 = tensor_x_position[i];
-        let x2 = tensor_x_position[j];
-        let new_x = f64::midpoint(x1, x2);
-        if let Some(parent_id) = contraction_tree.node(*i).parent_id() {
-            let y = tensor_position[&parent_id];
-            let color = tensor_color[&parent_id].clone();
-            dendogram_entries.push(DendogramEntry {
-                id: parent_id,
-                x: new_x,
-                y,
-                cost: y,
-                color,
-                children: Some((*i, *j)),
-            });
-            tensor_x_position.insert_new(parent_id, new_x);
-        }
-    }
-
-    let width_scaling = width / num_leaf_nodes as f64;
-    let height_scaling = height / dendogram_entries.last().unwrap().cost;
-
-    for dendogram_entry in &mut dendogram_entries {
-        dendogram_entry.x *= width_scaling;
-        dendogram_entry.y *= height_scaling;
-    }
[…]
