+++ tnc/Cargo.toml
@@ -29,9 +29,7 @@ serde = { version = "1.0.218", features = ["derive"] }
 postcard = { version = "1.1.3", default-features = false, features = ["use-std"] }
 log = { version = "0.4.22", features = ["kv_std", "kv_serde"] }
 rustc-hash = "2.1.1"
-chrono = "0.4.38"
 serde_json = "1.0.133"
-regex = "1.11.1"
 ndarray = "0.17.1"
 genetic_algorithm = { git = "https://github.com/Ectras/genetic-algorithm.git", branch = "f64_score" }
 rayon = "1.10.0"
