### 2026-01-23 8f235f9
Update tensor documentation
