### 2026-01-23 1c819b9
Bump mpi from 0.8.0 to 0.8.1 for benchmark
