### 2026-01-27 qc-tum/TNC#53 @Luthaf
Consider updating to hdf5-metno v0.12
This is part of https://github.com/openjournals/joss-reviews/issues/9598

This crate relies on hdf5-metno v0.11

https://github.com/qc-tum/TNC/blob/1c819b9b8b94eb79a3f7a58cf78e81cb7db3a681/tnc/Cargo.toml#L19

but this version does not know how to handle HDF5 v2.0. Updating to hdf5-metno v0.12 fixes the issue locally, but I'm not sure if this could have other unintended consequences.

```
  process didn't exit successfully: `~/code/tests/TNC/target/debug/build/hdf5-metno-sys-0de09f1984883029/build-script-build` (exit status: 101)
  --- stdout
  Attempting to find HDF5 via Homebrew (any version)...
  config.inc_dir = None
  Found Homebrew HDF5 headers at:
      "/opt/homebrew/opt/hdf5-mpi/include"
  Parsing HDF5 config from:
      "/opt/homebrew/opt/hdf5-mpi/include/H5pubconf.h"

  --- stderr

  thread 'main' (2376360) panicked at ~/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/hdf5-metno-sys-0.10.1/build.rs:220:21:
  Invalid H5_VERSION: "2.0.0"
```
