### 2026-02-01 a674c92
Bump taiki-e/install-action from 2.65.10 to 2.67.18
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.65.10 to 2.67.18.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/e0db384ad69f5ba2c6dd0129d8934e0d0ba465c1...650c5ca14212efbbf3e580844b04bdccf68dac31)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.67.18
  dependency-type: direct:production
  update-type: v
[…]
