### 2026-02-02 e661542
Move mpi broadcast test to integration tests
