+++ tnc/src/qasm/.antlr/Qasm2Parser.java
@@ -1,1513 +1,0 @@
-// Generated from /work_fast/ga87com/QuantumResearch/TensorStuff/GITHUB/tensornetworkcontractions/src/qasm/Qasm2Parser.g4 by ANTLR 4.13.1
-import org.antlr.v4.runtime.atn.*;
-import org.antlr.v4.runtime.dfa.DFA;
-import org.antlr.v4.runtime.*;
-import org.antlr.v4.runtime.misc.*;
-import org.antlr.v4.runtime.tree.*;
-import java.util.List;
-import java.util.Iterator;
-import java.util.ArrayList;
-
-@SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast", "CheckReturnValue"})
-public class Qasm2Parser extends Parser {
-	static { RuntimeMetaData.checkVersion("4.13.1", RuntimeMetaData.VERSION); }
-
-	protected static final DFA[] _decisionToDFA;
-	protected static final PredictionContextCache _sharedContextCache =
-		new PredictionContextCache();
-	public static final int
-		OPENQASM=1, INCLUDE=2, QREG=3, CREG=4, GATE=5, OPAQUE=6, RESET=7, MEASURE=8, 
-		BARRIER=9, IF=10, PI=11, U=12, CX=13, LBRACKET=14, RBRACKET=15, LBRACE=16, 
-		RBRACE=17, LPAREN=18, RPAREN=19, SEMICOLON=20, COMMA=21, DOT=22, ARROW=23, 
-		EQUALS=24, PLUS=25, MINUS=26, ASTERISK=27, SLASH=28, CARET=29, SIN=30, 
-		COS=31, TAN=32, EXP=33, LN=34, SQRT=35, Integer=36, Float=37, StringLiteral=38, 
-		Whitespace=39, Newline=40, LineComment=41, BlockComment=42, Identifier=43, 
-		VERSION_IDENTIFER_WHITESPACE=44, VersionSpecifier=45;
-	public static final int
-		RULE_program = 0, RULE_version = 1, RULE_statement = 2, RULE_includeStatement = 3, 
-		RULE_declaration = 4, RULE_gateDeclaration = 5, RULE_quantumOperation = 6, 
-		RULE_ifStatement = 7, RULE_barrier = 8, RULE_bodyStatement = 9, RULE_gateCall = 10, 
-		RULE_idlist = 11, RULE_mixedlist = 12, RULE_argument = 13, RULE_designator = 14, 
-		RULE_explist = 15, RULE_exp = 16;
-	private static String[] makeRuleNames() {
-		return new String[] {
-			"program", "version", "statement", "includeStatement", "declaration", 
-			"gateDeclaration", "quantumOperation", "ifStatement", "barrier", "bodyStatement", 
-			"gateCall", "idlist", "mixedlist", "argument", "designator", "explist", 
-			"exp"
-		};
-	}
-	public static final String[] ruleNames = makeRuleNames();
-
-	private static String[] makeLiteralNames() {
-		return new String[] {
-			null, "'OPENQASM'", "'include'", "'qreg'", "'creg'", "'gate'", "'opaque'", 
-			"'reset'", "'measure'", "'barrier'", "'if'", "'pi'", "'U'", "'CX'", "'['", 
-			"']'", "'{'", "'}'", "'('", "')'", "';'", "','", "'.'", "'->'", "'=='", 
-			"'+'", "'-'", "'*'", "'/'", "'^'", "'sin'", "'cos'", "'tan'", "'exp'", 
-			"'ln'", "'sqrt'"
-		};
-	}
-	private static final String[] _LITERAL_NAMES = makeLiteralNames();
[…]
