### 2026-02-09 0f55dbf
Use apt update before install in CI
