+++ .github/workflows/check.yml
@@ -40,7 +40,9 @@ jobs:
     name: clippy (stable)
     steps:
       - name: Install system requirements
-        run: sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
         with:
           submodules: true
@@ -73,7 +75,9 @@ jobs:
     name: doc (nightly)
     steps:
       - name: Install system requirements
-        run: sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
         with:
           submodules: true
@@ -92,7 +96,9 @@ jobs:
     name: feature-combinations (stable)
     steps:
       - name: Install system requirements
-        run: sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
         with:
           submodules: true
@@ -119,7 +125,9 @@ jobs:
     name: msrv (${{ matrix.msrv }})
     steps:
       - name: Install system requirements
-        run: sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev openmpi-bin libopenmpi-dev libboost-program-options-dev
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
         with:
           submodules: true
