### 2026-02-10 6ff58a3
Remove code coverage CI for benchmark crate
