### 2026-02-17 qc-tum/TNC#62 @Luthaf
Can not compile on macOS, trying to link to both libstdc++ and libc++
Done as part of https://github.com/openjournals/joss-reviews/issues/9598

----

I'm not sure where this comes from, and if this project is even supposed to run on macOS, but I now get the following error when running cargo test:

```
$ cargo test
   Compiling tnc v1.0.0 (~/code/TNC/tnc)
   Compiling jemalloc-sys v0.5.4+5.3.0-patched
   Compiling jemallocator v0.5.4
   Compiling benchmark v0.1.0 (~/code/TNC/benchmark)
error: linking with `cc` failed: exit status: 1
  |
  = note:  "cc" "/var/folders/nz/mgmqnhrj4mzfjlnqbgn02t780000gn/T/rustcBkCLfM/symbols.o" "<31 object files omitted>" "~/code/TNC/target/debug/deps/{libtnc-176d3c6edfac1b21,libpostcard-080861d894f02792,libcobs-db65f6345affb9ad,libkahypar-749052aeac6bbdf3,libkahypar_sys-b7d8ddd06f0cf15e,libantlr_rust-44b7dfce8bd14552,libuuid-777f5efee74fa67b,libtyped_arena-e081423e51f4149c,libbit_set-89e5319e5643c447,libbit_vec-7b41dcb3018186ee,libmurmur3-710ff607553b631f,libbetter_any-c41409faef8f1598,libparking_lot-a23410b9e415ce5d,libparking_lot_core-1e9fcb3eb13cfb64,libinstant-ee43f851e3c961fd,liblazy_static-fd8d1386ed5b7782,libhdf5_metno-9db326979c6f0169,libbitflags-306a78e11b8d65da,libndarray-9beaea39bb5fac37,libmatrixmultiply-580aa895a8c92f38,librawpointer-9905504089dcc385,libhdf5_metno_types-fb9642d9eb0e6cb2,libhdf5_metno_sys-89f6596a0e03ac15,libparking_lot-bf24967e5e444c2a,libparking_lot_core-dd59456d56ae6ee2,liblock_api-46e6444b1b674860,libscopeguard-ca1ff5b34f26b49f,libascii-57e0daf921ee473c,libgenetic_algorithm-e4c038698a098ee3,libstats-74817b3f50d1c240,libcardinality_estimator-8dc6a30b779b416f,libwyhash-cb3ddca34717b329,libfactorial-9ae6b0a6719e0a6a,libprimal_sieve-faeea0801c33b9b8,libprimal_estimate-f7902ac9ddf6f2f1,libprimal_bit-823e49f8a4b64a84,libhamming-de976455436776cb,libthread_local-31dd4d1457e9508d,librayon-1578b38f019cce83,librayon_core-4d282af7dfce5313,libcrossbeam_deque-59679f5fc1bbd1c4,libcrossbeam_epoch-3558c763a0fe4e52,libcrossbeam_utils-0c844a3b8af9ff18,libnum-9736b18afeb827ff,libnum_iter-004d0cc6007b1659,libnum_rational-ab65369f5cd431b8,libnohash_hasher-aa96b5b31cac452b,liblru-dd152e8d45e4f4c0,libhashbrown-801b05ec3f25a37a,libfoldhash-81285fc086c28929,libequivalent-d416f79c623eea2f,liballocator_api2-dff6215deafeda4b,libitertools-2dbab7e73df6b3b7,librand-049fb9dedda92829,librand_chacha-22452fe5783bfb90,librand_core-14031fcffb09e5f5,libgetrandom-eb96a8a0850a1c31,libfixedbitset-f9bd49d3a5958e83,librustengra-c661ff5d31a78ed3,libserde_pickle-ab9e6018597d3525,libnum_bigint-5f8777a9f11a2539,libnum_integer-b9fbf8bc5293f141,libiter_read-cf6072a85f0eae99,libbyteorder-de6fed5f1880a7fe,libcotengrust-319947fa0287c434,libpyo3-7d4df0215f0f7767,libpyo3_ffi-91c7f0e3e78784d9,libunindent-e4dc2a429c5cbcaf,libordered_float-2c7e3ff086d67106,libbit_set-6a1aa8d257cbecdd,libbit_vec-c8d641a28b914f52,liblog-90645142fb3393f5,libvalue_bag-ffe26feb418935d9,libvalue_bag_serde1-fb6a9c4041697c9d,libserde_fmt-e4e12aeb2486f4ef,liberased_serde-b1e1cbc0d3518035,libtypeid-6bb98ad5eaffb603,libtetra-93a86f9957a11a32,libbytemuck-0a0b61d51f2b8804,libcblas_sys-d323205550140860,libhptt-37fb71589a4eb88b,libhptt_sys-c3f9fd4deea3b204,libopenmp_sys-648557c960a68710,libfloat_cmp-b6c2c8dcd93e1f63,libintel_mkl_src-0851fc0c9bc82525,librustc_hash-0185af711e0f0fcd,librand-bfee4bbf70981329,librand_chacha-194c66ac8bfeb5aa,libppv_lite86-a3e2ab41e3b30f4d,libzerocopy-8f484ba22e6dc88f,librand_core-9c100d3b54fa9894,libgetrandom-802211f031548950,libcfg_if-83327022ead89bfc,libpermutation-74009e3ff924a892,libitertools-bfca84210d34bd8b,libeither-634090f7851035f6,libmpi-cec4ff8ad76784fa,libthiserror-e45d36c640f8142c,libsmallvec-8ac2f0c432b9b5d3,libmemoffset-f32a468025f09b98,libnum_complex-faa86e7b25adce93,libserde-9cce15e65e6dc919,libserde_core-243e8b0f7de16690,libnum_traits-b9e6569862afa6c2,liblibffi-96aff8fb98870b2f,liblibc-49b90cbd85e90172,liblibffi_sys-e0dd9695ba6b005b,libconv-ca2fec85bb4591d1,libcustom_derive-e74976610979d2c2,libonce_cell-4437ecc045f1a235,libmpi_sys-4d5dc3d1c11d1288}.rlib" "<sysroot>/lib/rustlib/aarch64-apple-darwin/lib/{libstd-*,libpanic_unwind-*,libobject-*,libmemchr-*,libaddr2line-*,libgimli-*,libcfg_if-*,librustc_demangle-*,libstd_detect-*,libhashbrown-*,librustc_std_workspace_alloc-*,libminiz_oxide-*,libadler2-*,libunwind-*,liblibc-*,librustc_std_workspace_core-*,liballoc-*,libcore-*,libcompiler_builtins-*}.rlib" "-lkahypar" "-lboost_program_options" "-lstdc++" "-lhdf5" "-lpython3.14" "-lhptt" "-lc++" "-liconv" "-lmpi" "-lSystem" "-lc" "-lm" "-arch" "arm64" "-mmacosx-version-min=11.0.0" "-L" "/opt/homebrew/Cellar/boost/1.90.0_1/lib" "-L" "~/code/TNC/target/debug/build/kahypar-sys-db6dff222d718fe4/out/lib" "-L" "~/code/TNC/target/debug/build/kahypar-sys-db6dff222d718fe4/out/lib64" "-L" "~/code/TNC/target/debug/build/libffi-sys-78ccd4988998ee55/out/libffi-root/lib" "-L" "~/code/TNC/target/debug/build/libffi-sys-78ccd4988998ee55/out/libffi-root/lib32" "-L" "~/code/TNC/target/debug/build/libffi-sys-78ccd4988998ee55/out/libffi-root/lib64" "-L" "~/code/TNC/target/debug/build/mpi-sys-4449ca8523296161/out" "-L" "~/code/TNC/target/debug/build/hptt-sys-305c76e6f04aac75/out/lib" "-L" "/opt/homebrew/opt/python@3.14/Frameworks/Python.framework/Versions/3.14/lib" "-L" "/opt/homebrew/opt/hdf5-mpi/lib" "-L" "/opt/homebrew/opt/hdf5-mpi/bin" "-L" "/opt/homebrew/Cellar/open-mpi/5.0.9/lib" "-o" "~/code/TNC/target/debug/examples/basic_usage-6ee395b1819c0cb4" "-Wl,-dead_strip" "-nodefaultlibs"
  = note: some arguments are omitted. use `--verbose` to show all linker arguments
  = note: ld: warning: search path '~/code/TNC/target/debug/build/kahypar-sys-db6dff222d718fe4/out/lib64' not found
          ld: warning: search path '~/code/TNC/target/debug/build/libffi-sys-78ccd4988998ee55/out/libffi-root/lib32' not found
          ld: warning: search path '~/code/TNC/target/debug/build/libffi-sys-78ccd4988998ee55/out/libffi-root/lib64' not found
          ld: library 'stdc++' not found
          clang: error: linker command failed with exit code 1 (use -v to see invocation)
```

notice in particular how the linker is trying to find both the LLVM C++ standard library (`lc++`) and the GNU one (`-lstdc++`). I'll continue checking this project on linux in the mean time.
