### 2026-02-17 qc-tum/TNC#63 @Luthaf
Tests are failling on linux, can not load `libpython.so`
I get the following error when running `cargo test`:

```
$ cargo test
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.15s
     Running unittests src/main.rs (target/debug/deps/benchmark-c4c14fdea5c7edd5)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests src/lib.rs (target/debug/deps/tnc-925732b71ed923a0)
~/code/TNC/target/debug/deps/tnc-925732b71ed923a0: error while loading shared libraries: libpython3.11.so.1.0: cannot open shared object file: No such file or directory
error: test failed, to rerun pass `-p tnc --lib`

Caused by:
  process didn't exit successfully: `~/code/TNC/target/debug/deps/tnc-925732b71ed923a0` (exit status: 127)
note: test exited abnormally; to see the full output pass --no-capture to the harness.
```

Not sure what is causing it, let me know if I can run some commands to help debug it!

----

Part of https://github.com/openjournals/joss-reviews/issues/9598
