### 2026-02-17 c670c35
Move hyperoptimization behind cotengra feature
