### 2026-02-17 19e9194
Add CI for MacOS
