### 2026-02-17 90af499
Use brew for macOS CI
