### 2026-02-17 720f8ed
Specify Boost path in macOS CI
