### 2026-02-18 qc-tum/TNC#63 @Luthaf
I do have Python3 installed, both from the system and from miniconda. Thanks for the fix, I'll check again without features enabled and then with cotengra as well!
