### 2026-02-18 qc-tum/TNC#62 @Ectras
That was another hardcoded linker flag in one of our C++ wrapper crates ([kahypar](https://github.com/Ectras/rust-kahypar)). It should be fixed by #66.

I'm sorry you're encountering so many build issues. We only targeted our own hardware when writing the code originally. But macOS and ARM support should be possible and within scope. I'll try to set up a macOS CI to find further build issues.
