+++ tnc/Cargo.toml
@@ -32,7 +32,7 @@ rustc-hash = "2.1.1"
 serde_json = "1.0.133"
 ndarray = "0.17.1"
 genetic_algorithm = { git = "https://github.com/Ectras/genetic-algorithm.git", branch = "f64_score" }
-rayon = "1.10.0"
+rayon = "1.11.0"
 ordered-float = "5.1.0"
 rustengra = { git = "https://github.com/Ectras/rustengra.git", optional = true }
 cotengrust = { git = "https://github.com/Ectras/cotengrust.git", branch = "rust_extension" }
