### 2026-02-19 361291e
Exclude mkl feature from macOS CI
