### 2026-02-19 8f4da9c
Allow MPI oversubscribe for MacOS CI
