+++ .github/workflows/test.yml
@@ -116,6 +116,7 @@ jobs:
     env:
       HDF5_DIR: /opt/homebrew/opt/hdf5
       BOOST_DIR: /opt/homebrew/opt/boost
+      PRTE_MCA_rmaps_default_mapping_policy: ":oversubscribe"
     steps:
       # if your project needs OpenSSL, uncomment this to fix Windows builds.
       # it's commented out by default as the install command takes 5-10m.
