### 2026-02-19 qc-tum/TNC#63 @Ectras
`cotengrust` is a Rust port of some cotengra functions. It is included in TNC by default. With the feature flag `cotengra`, we enable functions where we call the Python package cotengra. `pyo3` is a dependency of `cotengrust` (not because it calls Python, but because it provides a Python interface)

---

I haven't encountered this problem on my PC or in the CI, so I'm not sure yet what the issue could be. Maybe you could try explicitly setting the `PYTHONHOME` and `PYTHONPATH` as per this comment: https://github.com/PyO3/pyo3/discussions/3726#discussioncomment-14027946
