### 2026-02-19 qc-tum/TNC#69 @emapuljak
Improve docs and readme
Done as part of review https://github.com/openjournals/joss-reviews/issues/9598

Documentation is somewhat confusing.
Also README would benefit from restructuring.

I would add a high-level overview on the landing page of docs, explaining the modules' architectural diagram, and also connect tensor networks, contraction paths, partitioning, MPI.

Other suggestions:
1) Add installation guide (for now its only in readme)
1) Reorganize the sidebar into groups.
2) Add a quickstart (tutorial) guide.
3) enhance visibility of examples (not visible now).
4) Maybe it would be useful to explain MPI usage in the docs, just in case someone who is not familiar can get a small intro.
5) add community guidelines (contributing) to docs and readme
