### 2026-02-19 qc-tum/TNC#70 @emapuljak
Improve paper.md
Done as part of review https://github.com/openjournals/joss-reviews/issues/9598

1) Extend the summary with a bit better explanation suitable for non-specialist audience. If you are lacking of word count, then just try to rephrase this summary.
2) Missing references at the end of `paper.md`
3) Are you sure you covered all existing libraries? for example: https://github.com/jjacobx/qtnh
