### 2026-02-20 qc-tum/TNC#63 @Luthaf
More debugging from macOS:

>  Seems like it's not searching / picking up the correct path?

The code does directly link to libpython

```
$ otool -L /Users/guillaume/code/tests/TNC/target/debug/deps/tnc-098baf928ba1568c
/Users/guillaume/code/tests/TNC/target/debug/deps/tnc-098baf928ba1568c:
	/opt/homebrew/opt/boost/lib/libboost_program_options.dylib (compatibility version 0.0.0, current version 0.0.0)
	@rpath/libc++.1.dylib (compatibility version 1.0.0, current version 1.0.0)
	/opt/homebrew/opt/open-mpi/lib/libmpi.40.dylib (compatibility version 81.0.0, current version 81.7.0)
	/opt/homebrew/opt/hdf5-mpi/lib/libhdf5.320.dylib (compatibility version 320.0.0, current version 320.0.0)
	@rpath/libpython3.12.dylib (compatibility version 3.12.0, current version 3.12.0)
	@rpath/libiconv.2.dylib (compatibility version 10.0.0, current version 10.0.0)
	/usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1351.0.0)
```

But does not contain any rpath to load it

```
$ otool -l /Users/guillaume/code/tests/TNC/target/debug/deps/tnc-098baf928ba1568c | grep -i rpath
         name @rpath/libc++.1.dylib (offset 24)
         name @rpath/libpython3.12.dylib (offset 24)
         name @rpath/libiconv.2.dylib (offset 24)
```

-----

I can set `DYLD_FALLBACK_LIBRARY_PATH` to my miniconda installation, and then it finds libpython, but still fails some tests

```
$ cargo test --features=cotengra
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.10s
     Running unittests src/lib.rs (target/debug/deps/tnc-098baf928ba1568c)

running 164 tests
[...]

failures:

---- contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_complex stdout ----

thread 'contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_complex' (15654590) panicked at tnc/src/contractionpath/paths/tree_annealing.rs:34:9:
assertion failed: cotengra_check().is_ok()

---- contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_simple stdout ----

thread 'contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_simple' (15654593) panicked at tnc/src/contractionpath/paths/tree_annealing.rs:34:9:
assertion failed: cotengra_check().is_ok()

---- contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_complex stdout ----

thread 'contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_complex' (15654594) panicked at tnc/src/contractionpath/paths/tree_reconfiguration.rs:31:9:
assertion failed: cotengra_check().is_ok()

---- contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_simple stdout ----

thread 'contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_simple' (15654595) panicked at tnc/src/contractionpath/paths/tree_reconfiguration.rs:31:9:
assertion failed: cotengra_check().is_ok()

failures:
    contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_complex
    contractionpath::paths::tree_annealing::tests::test_anneal_tree_contract_order_simple
    contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_complex
    contractionpath::paths::tree_reconfiguration::tests::test_tree_contract_order_simple

test result: FAILED. 158 passed; 4 failed; 2 ignored; 0 measured; 0 filtered out; finished in 4.25s
```

I have a virtualenv with cotengra activated, and the code seems to find it

```
$ pip list
Package           Version
----------------- -------
alembic           1.18.4
autoray           0.8.4
colorlog          6.10.1
cotengra          0.7.5
kahypar           1.3.6
Mako              1.3.10
MarkupSafe        3.0.3
numpy             2.4.2
optuna            4.7.0
packaging         26.0
pip               25.0.1
PyYAML            6.0.3
SQLAlchemy        2.0.46
tqdm              4.67.3
typing_extensions 4.15.0
```
