### 2026-02-24 783586a
Replace basic_usage example by distributed_contraction
