### 2026-02-24 cd833ae
Use Circuit builder for sycamore_circuits
