### 2026-02-24 7f5c88f
Use sycamore circuit in distributed_contraction example
This makes it a bit more interesting, and is an easier way to create a
larger circuit (as opposed to putting a large QASM string in the file).
