### 2026-02-24 b68aabc
Add description of examples
