### 2026-02-25 a572561
Fix typo in CONTRIBUTING
