+++ tnc/src/_tutorial.rs
@@ -9,5 +9,6 @@
 //! and therefore run the risk of quickly becoming outdated.
 pub mod constructing_tns;
 pub mod hpc_overview;
+pub mod partitioning;
 pub mod pathfinding_and_contraction;
 pub mod tensors_structure;
