### 2026-03-01 d97f105
Bump taiki-e/install-action from 2.67.18 to 2.68.15
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.67.18 to 2.68.15.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/650c5ca14212efbbf3e580844b04bdccf68dac31...68675c5a5f1a6950c3975d33f3ae0ef155e5bf3d)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.68.15
  dependency-type: direct:production
  update-type: v
[…]
