### 2026-03-02 ea0de8e
Fix panic when running with 1 MPI rank
While it doesn't make much sense to run a distributed contraction with
only one rank, there's no reason we shouldn't allow this. The only
problem was that the tensor mapping was built depending on the toplevel
path, which is empty in such cases. Using {0:0} as mapping in such cases
instead avoids the problem.
