### 2026-03-02 b775c7e
Update example documentation
