+++ tnc/examples/README.md
@@ -10,8 +10,8 @@ The necessary features (if any) will be automatically activated.
 ## Overview
 Here is an overview of the examples in this folder:
 
-`distributed_contraction`:
-Distributed contraction of partitioned tensor networks using MPI.
+### `distributed_contraction`
+Distributed contraction of partitioned tensor networks using MPI. Execute with `mpirun` to actually run multiple processes (see the tutorial on running on HPC).
 - Creates a small [Sycamore](https://arxiv.org/pdf/1910.11333)-like circuit
 - Creates a tensor network that computes a single amplitude
 - Partitions the tensor network into multiple tensor networks
@@ -20,7 +20,7 @@ Distributed contraction of partitioned tensor networks using MPI.
 - Contracts the tensor networks on each compute node (intra-node)
 - Contracts the results (inter-node) to get the final result
 
-`local_contraction`:
+### `local_contraction`
 Local contraction of a tensor network on one machine.
 - Loads a small circuit (QASM code)
 - Creates a tensor network that computes the state vector
