+++ tnc/src/_tutorial/running_on_hpc.rs
@@ -19,9 +19,9 @@
 //! the steps to running this library are similar to:
 //! 1. Load modules for the required dependencies (HDF5, Boost, Python, ...)
 //! 2. Optionally create a virtual Python environment and install the Python
-//! dependencies for cotengra
+//!    dependencies for cotengra
 //! 3. Build the code using `cargo build -r` (assuming the login nodes have the same
-//! architecture as the compute nodes)
+//!    architecture as the compute nodes)
 //! 4. Write a SLURM script that calls `mpirun` on the compiled binary
 //! 5. Submit the job
 
