+++ tnc/examples/repartitioning.rs
@@ -1,0 +1,140 @@
+use std::time::Duration;
+
+use rand::{rngs::StdRng, SeedableRng};
+use tnc::{
+    builders::sycamore_circuit::sycamore_circuit,
+    contractionpath::{
+        communication_schemes::CommunicationScheme,
+        contraction_cost::{communication_path_cost, contract_path_cost},
+        paths::{
+            cotengrust::{Cotengrust, OptMethod},
+            FindPath,
+        },
+        repartitioning::simulated_annealing::{self, IntermediatePartitioningModel},
+    },
+    tensornetwork::{
+        partitioning::{
+            find_partitioning, partition_config::PartitioningStrategy, partition_tensor_network,
+        },
+        tensor::Tensor,
+    },
+};
+
+fn compute_cost(tensor: &Tensor, partitioning: &[usize]) -> f64 {
+    // Partition the tensor
+    let partitioned_tn = partition_tensor_network(tensor.clone(), partitioning);
+
+    // Find a contraction path
+    let mut opt = Cotengrust::new(&partitioned_tn, OptMethod::Greedy);
+    opt.find_path();
+    let path = opt.get_best_replace_path();
+
+    // Get the serial contraction cost of each partition
+    let cost_per_partition = partitioned_tn
+        .tensors()
+        .iter()
+        .enumerate()
+        .map(|(i, t)| {
+            let (ops, _mem) = contract_path_cost(t.tensors(), &path.nested[&i], true);
+            ops
+        })
+        .collect::<Vec<_>>();
+
+    // Get the shape of the partition tensors (i.e., the tensors resulting from the contraction of the respective partition)
+    let partition_tensors = partitioned_tn
+        .tensors()
+        .iter()
+        .map(Tensor::external_tensor)
+        .collect::<Vec<_>>();
+
+    // Get the contraction cost of contracting the final partition tensors in parallel
+    let (final_cost, _) = communication_path_cost(
+        &partition_tensors,
+        &path.toplevel,
+        true,
+        true,
+        Some(&cost_per_partition),
+    );
+    final_cost
+}
+
+fn rebalance(
+    tensor: &Tensor,
+    initial_partitioning: &[usize],
+    rng: &mut StdRng,
+    communication_scheme: CommunicationScheme,
+) -> Vec<usize> {
+    // Unfortunately, the simulated annealing algorithm needs a few things:
+    // The partition tensors and some initial contraction paths for each partition.
+    // Here, we recompute these things, but in practice, you might already have them
+    // (for example, from the cost computation of the naive partitioning).
+
+    // Partition the tensor
+    let partitioned_tn = partition_tensor_network(tensor.clone(), initial_partitioning);
+
+    // Get the partition tensors
+    let partition_tensors = partitioned_tn
+        .tensors()
+        .iter()
+        .map(Tensor::external_tensor)
+        .collect::<Vec<_>>();
+
+    // Find a contraction path for the partitioned tensor network
+    let mut opt = Cotengrust::new(&partitioned_tn, OptMethod::Greedy);
+    opt.find_path();
+    let path = opt.get_best_replace_path();
+
+    // Collect the contractions path for each partition
+    let initial_contraction_paths = (0..partition_tensors.len())
+        .map(|i| path.nested[&i].clone().into_simple())
+        .collect::<Vec<_>>();
+
+    // Run the rebalancing
+    let (solution, _) = simulated_annealing::balance_partitions(
+        IntermediatePartitioningModel {
+            tensor,
+            communication_scheme,
+            memory_limit: None,
+        },
+        (
+            initial_partitioning.to_vec(),
+            partition_tensors,
[…]
