### 2026-03-02 56fd702
Remove useless lifetime from simulated annealing
