### 2026-03-02 eaa03c0
Compute memory in SA only when memory limit is set
