### 2026-03-02 a042330
Simplify source partition selection in SA
