### 2026-03-02 b447e2c
Add method to compute initial solution for SA
