### 2026-03-02 03bf458
Refactor SA evaluation into common method
