### 2026-03-02 e187f9c
Remove unnecessary explicit generic argument
