### 2026-03-02 46ba288
Avoid panic in is_connected for single tensor
