+++ tnc/src/tensornetwork/tensordata.rs
@@ -13,10 +13,9 @@ use crate::{
 /// The data of a tensor.
 #[derive(Default, Debug, Clone, Serialize, Deserialize)]
 pub enum TensorData {
-    /// This is for composite tensors that have not been contracted yet, as well as
-    /// empty tensors in general.
+    /// No data.
     #[default]
-    Uncontracted,
+    None,
     /// The data is loaded from a HDF5 file.
     File((PathBuf, bool)),
     /// A quantum gate. The name must be registered in the gates module.
@@ -39,7 +38,7 @@ impl TensorData {
     /// Consumes the tensor data and returns the contained tensor.
     pub fn into_data(self) -> DataTensor {
         match self {
-            TensorData::Uncontracted => panic!("Cannot convert uncontracted tensor to data"),
+            TensorData::None => panic!("Cannot convert uncontracted tensor to data"),
             TensorData::File((filename, adjoint)) => {
                 let mut data = load_data(filename).unwrap();
                 if adjoint {
@@ -61,7 +60,7 @@ impl TensorData {
     /// Returns the adjoint of this data.
     pub fn adjoint(self) -> Self {
         match self {
-            TensorData::Uncontracted => TensorData::Uncontracted,
+            TensorData::None => TensorData::None,
             TensorData::File((filename, adjoint)) => TensorData::File((filename, !adjoint)),
             TensorData::Gate((name, params, adjoint)) => TensorData::Gate((name, params, !adjoint)),
             TensorData::Matrix(mut tensor) => {
@@ -84,7 +83,7 @@ impl ApproxEq for &TensorData {
                 TensorData::Gate((name_r, angles_r, adjoint_r)),
             ) => name_l == name_r && adjoint_l == adjoint_r && angles_l.approx_eq(angles_r, margin),
             (TensorData::Matrix(l0), TensorData::Matrix(r0)) => l0.approx_eq(r0, margin),
-            (TensorData::Uncontracted, TensorData::Uncontracted) => true,
+            (TensorData::None, TensorData::None) => true,
             _ => false,
         }
     }
