+++ tnc/src/builders/circuit_builder.rs
@@ -7,7 +7,7 @@ use num_complex::Complex64;
 use permutation::Permutation;
 
 use crate::tensornetwork::{
-    tensor::{EdgeIndex, Tensor},
+    tensor::{CompositeTensor, EdgeIndex, LeafTensor},
     tensordata::TensorData,
 };
 
@@ -83,18 +83,12 @@ impl Permutor {
     }
 
     /// Permutates the tensor according to the stored permutation.
-    pub fn apply(&self, tensor: Tensor) -> Tensor {
-        assert!(tensor.is_leaf());
+    pub fn apply(&self, tensor: LeafTensor) -> LeafTensor {
         if self.is_identity() {
             return tensor;
         }
 
-        let Tensor {
-            mut legs,
-            mut bond_dims,
-            tensordata,
-            ..
-        } = tensor;
+        let (mut legs, mut bond_dims, tensordata) = tensor.into_inner();
         let mut data = tensordata.into_data();
 
         // Find the permutation
@@ -105,12 +99,7 @@ impl Permutor {
         perm.apply_slice_in_place(&mut bond_dims);
         data.transpose(&perm);
 
-        Tensor {
-            tensors: vec![],
-            legs,
-            bond_dims,
-            tensordata: TensorData::Matrix(data),
-        }
+        LeafTensor::new_with_data(legs, bond_dims, TensorData::Matrix(data))
     }
 
     /// Returns whether the permutor is identity, in which case it won't have any
@@ -137,8 +126,8 @@ pub struct Circuit {
     open_edges: Vec<EdgeIndex>,
     /// The next edge to be used.
     next_edge: usize,
-    /// The tensors representing the circuit.
-    tensors: Vec<Tensor>,
+    /// The tensor network representing the circuit.
+    tensor_network: CompositeTensor,
 }
 
 impl Circuit {
@@ -185,13 +174,13 @@ impl Circuit {
         let previous_qubits = self.num_qubits();
 
         self.open_edges.reserve(size);
-        self.tensors.reserve(size);
+        self.tensor_network.reserve(size);
         for _ in 0..size {
             let edge = self.new_edge();
             self.open_edges.push(edge);
-            let mut ket0 = Tensor::new_from_const(vec![edge], 2);
+            let mut ket0 = LeafTensor::new_from_const(vec![edge], 2);
             ket0.set_tensor_data(Self::ket0());
-            self.tensors.push(ket0);
+            self.tensor_network.push_tensor(ket0);
         }
 
         QuantumRegister {
@@ -220,11 +209,11 @@ impl Circuit {
         }
 
         // Create the new tensor
-        let mut new_tensor = Tensor::new_from_const(edges, 2);
+        let mut new_tensor = LeafTensor::new_from_const(edges, 2);
         new_tensor.set_tensor_data(gate);
 
         // Push the new tensor to the tensors
-        self.tensors.push(new_tensor);
+        self.tensor_network.push_tensor(new_tensor);
     }
 
     /// Converts the circuit to a tensor network that computes the amplitude for the
@@ -240,11 +229,11 @@ impl Circuit {
     /// natural order, i.e., sorted by increasing qubit number. If the bitstring
     /// contains no wildcards, the final result is a scalar and the permutator can be
     /// ignored.
-    pub fn into_amplitude_network(mut self, bitstring: &str) -> (Tensor, Permutor) {
+    pub fn into_amplitude_network(mut self, bitstring: &str) -> (CompositeTensor, Permutor) {
         assert_eq!(bitstring.len(), self.num_qubits());
 
         // Apply the final bras
-        self.tensors.reserve(bitstring.len());
+        self.tensor_network.reserve(bitstring.len());
         let mut final_legs = Vec::new();
[…]

[…]
