### 2026-03-06 5c25bf9
Use explicit structs for Composite and Leaf tensors
This makes intent clearer. For example, the contract method can now take
a composite tensor as input and return a leaf tensor. At the same time,
this also prevents bugs that were possible before, such as trying to get
the children of a leaf tensor or trying to get the data of a composite
tensor.
