+++ tnc/src/contractionpath/paths/cotengrust.rs
@@ -7,7 +7,7 @@ use rustc_hash::FxHashMap;
 use crate::contractionpath::contraction_cost::contract_path_cost;
 use crate::contractionpath::paths::FindPath;
 use crate::contractionpath::{ssa_replace_ordering, ContractionPath, SimplePath};
-use crate::tensornetwork::tensor::Tensor;
+use crate::tensornetwork::tensor::{CompositeTensor, LeafTensor, TensorType};
 
 /// The optimization method to use.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -23,7 +23,7 @@ pub enum OptMethod {
 /// A contraction path finder using the `cotengrust` library.
 #[derive(Debug, Clone)]
 pub struct Cotengrust<'a> {
-    tensor: &'a Tensor,
+    tensor: &'a CompositeTensor,
     best_path: ContractionPath,
     best_flops: f64,
     best_size: f64,
@@ -32,7 +32,7 @@ pub struct Cotengrust<'a> {
 
 impl<'a> Cotengrust<'a> {
     /// Creates a new Cotengrust optimizer using the specified optimization method.
-    pub fn new(tensor: &'a Tensor, opt_method: OptMethod) -> Self {
+    pub fn new(tensor: &'a CompositeTensor, opt_method: OptMethod) -> Self {
         Self {
             tensor,
             opt_method,
@@ -44,7 +44,7 @@ impl<'a> Cotengrust<'a> {
 
     /// Finds a contraction path for a "classical" tensor network, i.e. the inputs
     /// are all leaf tensors.
-    fn optimize_single(&self, inputs: &[Tensor], output: &Tensor) -> SimplePath {
+    fn optimize_single(&self, inputs: &[LeafTensor], output: &LeafTensor) -> SimplePath {
         // Check if the inputs are empty (cotengrust does not handle this gracefully)
         if inputs.is_empty() {
             return SimplePath::default();
@@ -100,8 +100,8 @@ impl<'a> Cotengrust<'a> {
 
 /// Converts tensor leg inputs to chars. Creates new inputs, outputs and size_dict that can be fed to Cotengra.
 fn tensor_legs_to_digit(
-    inputs: &[Tensor],
-    output: &Tensor,
+    inputs: &[LeafTensor],
+    output: &LeafTensor,
 ) -> (Vec<Vec<char>>, Vec<char>, FxHashMap<char, f32>) {
     fn leg_to_char(leg: usize) -> char {
         char::from_u32(leg.try_into().unwrap()).unwrap()
@@ -113,9 +113,9 @@ fn tensor_legs_to_digit(
     for (tensor, labels) in zip(inputs, new_inputs.iter_mut()) {
         labels.reserve_exact(tensor.legs().len());
         for (leg, dim) in tensor.edges() {
-            let character = leg_to_char(*leg);
+            let character = leg_to_char(leg);
             labels.push(character);
-            new_size_dict.insert(character, *dim as f32);
+            new_size_dict.insert(character, dim as f32);
         }
     }
     (new_inputs, new_output, new_size_dict)
@@ -125,19 +125,26 @@ impl FindPath for Cotengrust<'_> {
     fn find_path(&mut self) {
         // Handle nested tensors first
         let mut nested_paths = FxHashMap::default();
-        let mut inputs = self.tensor.tensors().clone();
-        for (index, input_tensor) in inputs.iter_mut().enumerate() {
-            if input_tensor.is_composite() {
-                let mut ct = Cotengrust::new(input_tensor, self.opt_method);
-                ct.find_path();
-                nested_paths.insert(index, ct.get_best_path().clone());
[…]
