+++ tnc/src/contractionpath/repartitioning/simulated_annealing.rs
@@ -22,7 +22,10 @@ use crate::{
         repartitioning::compute_solution,
         SimplePath,
     },
-    tensornetwork::{partitioning::partition_tensor_network, tensor::Tensor},
+    tensornetwork::{
+        partitioning::partition_tensor_network,
+        tensor::{CompositeTensor, LeafTensor},
+    },
 };
 
 type ScoreType = NotNan<f64>;
@@ -166,7 +169,7 @@ impl SimulatedAnnealingOptimizer {
 
 /// Common evaluation method for all simulated annealing methods.
 fn evaluate_partitioning<R>(
-    tensor: &Tensor,
+    tensor: &CompositeTensor,
     partitioning: &[usize],
     communication_scheme: CommunicationScheme,
     memory_limit: Option<f64>,
@@ -197,7 +200,7 @@ where
 
 /// A simulated annealing model that moves a random tensor between random partitions.
 pub struct NaivePartitioningModel<'a> {
-    pub tensor: &'a Tensor,
+    pub tensor: &'a CompositeTensor,
     pub num_partitions: usize,
     pub communication_scheme: CommunicationScheme,
     pub memory_limit: Option<f64>,
@@ -236,7 +239,7 @@ impl OptModel for NaivePartitioningModel<'_> {
 
 /// A simulated annealing model that moves a random subtree between random partitions.
 pub struct NaiveIntermediatePartitioningModel<'a> {
-    pub tensor: &'a Tensor,
+    pub tensor: &'a CompositeTensor,
     pub num_partitions: usize,
     pub communication_scheme: CommunicationScheme,
     pub memory_limit: Option<f64>,
@@ -312,8 +315,8 @@ impl OptModel for NaiveIntermediatePartitioningModel<'_> {
         }
 
         // Recompute the contraction path for both partitions
-        let mut from_tensor = Tensor::default();
-        let mut to_tensor = Tensor::default();
+        let mut from_tensor = CompositeTensor::default();
+        let mut to_tensor = CompositeTensor::default();
         for (partition_index, tensor) in zip(&partitioning, self.tensor.tensors()) {
             if *partition_index == source_partition {
                 from_tensor.push_tensor(tensor.clone());
@@ -349,13 +352,13 @@ impl OptModel for NaiveIntermediatePartitioningModel<'_> {
 /// A simulated annealing model that moves a random tensor to the partition that
 /// maximizes memory reduction.
 pub struct LeafPartitioningModel<'a> {
-    pub tensor: &'a Tensor,
+    pub tensor: &'a CompositeTensor,
     pub communication_scheme: CommunicationScheme,
     pub memory_limit: Option<f64>,
 }
 
 impl OptModel for LeafPartitioningModel<'_> {
-    type SolutionType = (Vec<usize>, Vec<Tensor>);
+    type SolutionType = (Vec<usize>, Vec<LeafTensor>);
 
     fn generate_trial_solution<R: Rng>(
         &self,
@@ -364,7 +367,7 @@ impl OptModel for LeafPartitioningModel<'_> {
     ) -> Self::SolutionType {
         let (mut partitioning, mut partition_tensors) = current_solution;
         let tensor_index = rng.random_range(0..partitioning.len());
-        let shifted_tensor = self.tensor.tensor(tensor_index);
+        let shifted_tensor = self.tensor.tensor(tensor_index).as_leaf().unwrap();
         let source_partition = partitioning[tensor_index];
 
         let (new_partition, _) = partition_tensors
@@ -404,7 +407,7 @@ impl OptModel for LeafPartitioningModel<'_> {
 /// A simulated annealing model that moves a random subtree to the partition that
 /// maximizes memory reduction.
[…]
