+++ tnc/src/contractionpath/paths/tree_reconfiguration.rs
@@ -8,7 +8,7 @@ use crate::{
         paths::{BasicContractionPathResult, CostType, Pathfinder},
         ssa_replace_ordering, ContractionPath,
     },
-    tensornetwork::tensor::Tensor,
+    tensornetwork::tensor::CompositeTensor,
 };
 
 /// Creates an interface to `rustengra` an interface to access `Cotengra` methods in
@@ -35,21 +35,22 @@ impl TreeReconfigure {
 impl Pathfinder for TreeReconfigure {
     type Result = BasicContractionPathResult;
 
-    fn find_path(&mut self, tensor: &Tensor) -> BasicContractionPathResult {
+    fn find_path(&mut self, tensor: &CompositeTensor) -> BasicContractionPathResult {
         // Map tensors to legs
         let inputs = tensor
             .tensors()
             .iter()
-            .map(|tensor| tensor.legs().clone())
+            .map(|tensor| tensor.as_leaf().unwrap().legs().clone())
             .collect_vec();
         let outputs = tensor.external_tensor();
-        let size_dict = tensor.tensors().iter().map(Tensor::edges).fold(
-            FxHashMap::default(),
-            |mut acc, edges| {
+        let size_dict = tensor
+            .tensors()
+            .iter()
+            .map(|tensor| tensor.as_leaf().unwrap().edges())
+            .fold(FxHashMap::default(), |mut acc, edges| {
                 acc.extend(edges);
                 acc
-            },
-        );
+            });
 
         let best_path =
             cotengra_optimized_greedy(&inputs, outputs.legs(), &size_dict, self.subtree_size)
@@ -77,20 +78,20 @@ mod tests {
     use crate::{
         contractionpath::paths::{CostType, Pathfinder},
         path,
-        tensornetwork::tensor::Tensor,
+        tensornetwork::tensor::LeafTensor,
     };
 
-    fn setup_simple() -> Tensor {
+    fn setup_simple() -> CompositeTensor {
         let bond_dims =
             FxHashMap::from_iter([(0, 5), (1, 2), (2, 6), (3, 8), (4, 1), (5, 3), (6, 4)]);
-        Tensor::new_composite(vec![
-            Tensor::new_from_map(vec![4, 3, 2], &bond_dims),
-            Tensor::new_from_map(vec![0, 1, 3, 2], &bond_dims),
-            Tensor::new_from_map(vec![4, 5, 6], &bond_dims),
+        CompositeTensor::new(vec![
+            LeafTensor::new_from_map(vec![4, 3, 2], &bond_dims),
+            LeafTensor::new_from_map(vec![0, 1, 3, 2], &bond_dims),
+            LeafTensor::new_from_map(vec![4, 5, 6], &bond_dims),
         ])
     }
 
-    fn setup_complex() -> Tensor {
+    fn setup_complex() -> CompositeTensor {
         let bond_dims = FxHashMap::from_iter([
             (0, 27),
             (1, 18),
@@ -105,13 +106,13 @@ mod tests {
             (10, 5),
             (11, 17),
         ]);
-        Tensor::new_composite(vec![
-            Tensor::new_from_map(vec![4, 3, 2], &bond_dims),
-            Tensor::new_from_map(vec![0, 1, 3, 2], &bond_dims),
-            Tensor::new_from_map(vec![4, 5, 6], &bond_dims),
-            Tensor::new_from_map(vec![6, 8, 9], &bond_dims),
-            Tensor::new_from_map(vec![10, 8, 9], &bond_dims),
-            Tensor::new_from_map(vec![5, 1, 0], &bond_dims),
+        CompositeTensor::new(vec![
[…]

[…]
