+++ paper/paper.bib
@@ -116,3 +116,69 @@
   primaryclass  = {physics.comp-ph},
   url           = {https://arxiv.org/abs/1402.0939}
 }
+
+@misc{MKL,
+  author       = {Intel},
+  howpublished = {\url{https://software.intel.com/en-us/intel-mkl}},
+  title        = {oneAPI Math Kernel Library (oneMKL)}
+}
+
+@article{ExaTN,
+  author  = {Lyakh, Dmitry I. and Nguyen, Thien and Claudino, Daniel and Dumitrescu, Eugene and McCaskey, Alexander J.},
+  journal = {Frontiers Appl. Math. Stat.},
+  title   = {{ExaTN}: {S}calable {GPU}-accelerated high-performance processing of general tensor networks at exascale},
+  year    = {2022},
+  pages   = {838601},
+  volume  = {8},
+  doi     = {10.3389/fams.2022.838601}
+}
+
+@article{Jet,
+  author  = {Vincent, Trevor and O'Riordan, Lee J. and Andrenkov, Mikhail and Brown, Jack and Killoran, Nathan and Qi, Haoyu and Dhand, Ish},
+  journal = {Quantum},
+  title   = {Jet: {F}ast quantum circuit simulations with parallel task-based tensor-network contraction},
+  year    = {2022},
+  issn    = {2521-327X},
+  month   = may,
+  pages   = {709},
+  volume  = {6},
+  doi     = {10.22331/q-2022-05-09-709}
+}
+
+@inproceedings{QuantEx,
+  author    = {Brennan, John and Allalen, Momme and Brayford, David and Hanley, Kenneth and Iapichino, Luigi and O{\textquoteright}Riordan, Lee J. and Doyle, Myles and Moran, Niall},
+  booktitle = {2021 IEEE/ACM Second International Workshop on Quantum Computing Software (QCS)},
+  title     = {Tensor Network Circuit Simulation at Exascale},
+  year      = {2021},
+  pages     = {20--26},
+  doi       = {10.1109/QCS54837.2021.00006},
+  keywords  = {Quantum system;Tensors;Circuit simulation;Conferences;Computational modeling;Software;Quantum circuit;quantum circuit;tensor network;exascale;HPC}
+}
+
+@inproceedings{QTensor,
+  author    = {{Lykov}, Danylo and {Alexeev}, Yuri and {Ibrahim}, Cameron and {Galda}, Alexey},
+  title     = {{QTensor: Fast QAOA Quantum Simulator}},
+  booktitle = {APS March Meeting Abstracts},
+  year      = 2021,
+  series    = {APS Meeting Abstracts},
+  volume    = {2021},
+  month     = {jan},
+  eid       = {M34.007},
+  pages     = {M34.007},
+  adsurl    = {https://ui.adsabs.harvard.edu/abs/2021APS..MARM34007L},
+  adsnote   = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{Garcia2024,
+  author        = {Garc{\'{i}}a, Marcos D{\'{i}}ez and Romero, Antonio M{\'{a}}rquez},
+  month         = aug,
+  title         = {Survey on Computational Applications of Tensor Network Simulations},
+  year          = {2024},
+  archiveprefix = {arXiv},
+  copyright     = {arXiv.org perpetual, non-exclusive license},
+  doi           = {10.48550/ARXIV.2408.05011},
+  eprint        = {2408.05011},
+  keywords      = {Quantum Physics (quant-ph), FOS: Physical sciences},
+  primaryclass  = {quant-ph},
+  publisher     = {arXiv}
+}
