+++ paper/paper.md
@@ -26,34 +26,49 @@ bibliography: paper.bib
 
 # Summary
 
-TNC is a Rust-based library for efficient contraction of tensor networks, a key technique for classical simulation of quantum circuits.
-Designed for distributed-memory environments common in high-performance computing (HPC), TNC partitions tensor networks to enable parallel contraction across multiple nodes.
-While optimized for quantum circuit simulation, TNC is general enough to handle arbitrary tensor networks, combining high performance with Rust's memory safety guarantees.
+Since real quantum computers continue to have high error rates, classical error-free simulation of circuits is essential for verification, benchmarking, and algorithm design.
+A commonly used tool for the simulation are tensor networks, which represent a quantum circuit as network of multi-dimensional arrays that must be multiplied ("contracted") to compute the result of the circuit.
+TNC is a Rust-based library for efficiently contracting large tensor networks.
+Designed for distributed-memory environments common in high-performance computing (HPC), TNC partitions tensor networks to enable parallel contraction across multiple compute nodes.
+While optimized for quantum circuit simulation, TNC is general enough to contract arbitrary tensor networks, combining high performance with Rust's memory safety guarantees.
 
 # Statement of need
 
 With the rapid advancements in quantum computing, ever larger quantum circuits are being explored, which require more computing resources.
-Since real quantum hardware is often not yet available and has to fight problems such as high error rates, research on efficient classical simulation methods is crucial to bridge this gap.
-Tensor networks have been shown to be a viable tool for classical simulation, disproving even two major claims of quantum advantage, i.e., the realization of an algorithm on real quantum hardware which would be infeasible to simulate on classical hardware [@Pan2022;@Patra2024].
+Real quantum hardware is often not yet available, however, has limited qubit counts, and has high error rates due to imperfect gates and noise induced by the environment.
+Hence, research on efficient classical simulation methods is crucial to bridge this gap.
+Tensor networks have been shown to be a viable tool for classical simulation, even disproving two major claims of quantum advantage, i.e., the realization of an algorithm on real quantum hardware which would be infeasible to simulate on classical hardware [@Pan2022;@Patra2024].
+Apart from quantum simulation, they are also used in other areas, such as machine learning, quantum optimization, quantum chemistry, and quantum error correction [@Garcia2024].
 
-However, existing libraries for tensor network contractions often only consider shared-memory parallelism [@Pfeifer2015], are not open-source [@Bayraktar2023], or are written in Python [@Gray2021] and hence less suited for HPC.
+However, existing libraries for tensor network contractions often only consider shared-memory parallelism [@Pfeifer2015], are not open-source [@Bayraktar2023], or are written in Python [@QTensor] and hence less suited for HPC.
 Furthermore, parallelization is usually done by a technique called *slicing*, which incurs computational overhead.
+Popular libraries using slicing include ExaTN [@ExaTN] and Jet [@Jet].
 TNC addresses these limitations by targeting distributed-memory systems, being fully open-source, implemented in Rust, and employing partitioning rather than slicing for parallelization.
+The use of Rust further provides strong guarantees of memory safety.
 
 # Features
 
 The user can construct quantum circuits using the library's API or import them from code written in the widely-used OpenQASM2 language [@Cross2017].
 From a given circuit, the user can then generate tensor networks that compute the expectation value, the amplitudes to specific measurement outcomes, or the full state vector.
-Furthermore, random tensor networks can be created as well, e.g., in the form of the original Sycamore experiment [@Arute2019].
+Furthermore, random tensor networks can be created as well, for instance, in the form of the original Sycamore experiment [@Arute2019].
+The library supports serialization of tensors and tensor networks to and from HDF5 files.
 
[…]
