### 2026-03-10 03c2764
Add testing book CI
