+++ .github/workflows/doc.yml
@@ -31,8 +31,8 @@ jobs:
         run: cargo doc --no-deps
       - name: Add redirect
         run: echo '<meta http-equiv="refresh" content="0;url=tnc/index.html">' > target/doc/index.html
-      - uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
-        name: Upload artifact
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
         id: deployment
         with:
           path: target/doc
