+++ .github/workflows/book.yml
@@ -1,0 +1,39 @@
+# Test CI to build the book
+permissions:
+  contents: read
+on:
+  push:
+concurrency:
+  group: pages
+  cancel-in-progress: true
+name: book
+jobs:
+  build-book:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
+        with:
+          submodules: true
+      - name: Setup mdBook
+        uses: peaceiris/actions-mdbook@ee69d230fe19748b7abf22df32acaa93833fad08 # tag=v2.0.0
+        with:
+          mdbook-version: 0.5.2
+      - name: Build book
+        run: mdbook build book
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
+        with:
+          path: book/book
+  deploy:
+    runs-on: ubuntu-latest
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    permissions:
+      pages: write
+      id-token: write
+    needs: build-book
+    steps:
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@d6db90164ac5ed86f2b6aed7e0febac5b3c0c03e # tag=v4.0.5
