### 2026-03-12 2659149
Add book CI
