+++ .github/workflows/doc.yml
@@ -1,7 +1,8 @@
 # This workflow runs for pushes to main. It runs the following jobs:
-# - build: builds the documentation
+# - build-doc: builds the documentation
+# - build-book: builds the book
 # - cleanup: cleans all previous deployments to GitHub Pages
-# - deploy: deploys the documentation to GitHub Pages
+# - deploy: deploys the documentation and book to GitHub Pages
 permissions:
   contents: read
 on:
@@ -12,9 +13,8 @@ concurrency:
   cancel-in-progress: true
 name: docs
 jobs:
-  build:
+  build-doc:
     runs-on: ubuntu-latest
-    name: build (stable)
     steps:
       - name: Install system requirements
         run: |
@@ -31,15 +31,51 @@ jobs:
         run: cargo doc --no-deps
       - name: Add redirect
         run: echo '<meta http-equiv="refresh" content="0;url=tnc/index.html">' > target/doc/index.html
-      - uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
-        name: Upload artifact
+      - name: Upload doc artifact
+        uses: actions/upload-artifact@bbbca2ddaa5d8feaa63e36b76fdaad77386f024f # tag=v7.0.0
+        with:
+          name: doc
+          path: target/doc
+  build-book:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
+        with:
+          submodules: true
+      - name: Setup mdBook
+        uses: peaceiris/actions-mdbook@ee69d230fe19748b7abf22df32acaa93833fad08 # tag=v2.0.0
+        with:
+          mdbook-version: 0.5.2
+      - name: Build book
+        run: mdbook build book
+      - name: Upload book artifact
+        uses: actions/upload-artifact@bbbca2ddaa5d8feaa63e36b76fdaad77386f024f # tag=v7.0.0
+        with:
+          name: book
+          path: book/book
+  upload-folder:
+    runs-on: ubuntu-latest
+    needs: [build-doc, build-book]
+    steps:
+      - name: Download doc artifact
+        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # tag=v8.0.1
+        with:
+          name: doc
+          path: target/pages/doc
+      - name: Download book artifact
+        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # tag=v8.0.1
+        with:
+          name: book
+          path: target/pages/book
+      - name: Upload artifacts folder
+        uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
         id: deployment
         with:
-          path: target/doc
+          path: target/pages
   cleanup:
     runs-on: ubuntu-latest
     permissions: write-all
-    needs: build
+    needs: upload-folder
     steps:
       - name: Delete previous deployments
         uses: strumwolf/delete-deployment-environment@a4825dd9648c57da8437a4885c3fcad58beac69c # tag=v3.0.0
