### 2026-03-12 3794778
Bump taiki-e/install-action from 2.68.15 to 2.68.26
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.68.15 to 2.68.26.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/68675c5a5f1a6950c3975d33f3ae0ef155e5bf3d...64c5c20c872907b6f7cd50994ac189e7274160f2)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.68.26
  dependency-type: direct:production
  update-type: v
[…]
