+++ .github/workflows/test.yml
@@ -172,7 +172,7 @@ jobs:
           toolchain: nightly
           components: llvm-tools-preview
       - name: cargo install cargo-llvm-cov
-        uses: taiki-e/install-action@68675c5a5f1a6950c3975d33f3ae0ef155e5bf3d # tag=v2.68.15
+        uses: taiki-e/install-action@64c5c20c872907b6f7cd50994ac189e7274160f2 # tag=v2.68.26
         with:
           tool: cargo-llvm-cov
       - name: cargo generate-lockfile
