### 2026-03-12 831691d
Fix links from book to documentation
