+++ book/src/pathfinding_and_contraction.md
@@ -1,3 +1,5 @@
+{{#include links.md}}
+
 # Pathfinding and Contraction
 
 ## Pathfinding
