+++ book/src/running_on_hpc.md
@@ -1,3 +1,5 @@
+{{#include links.md}}
+
 # Running on HPC
 
 ## Running with MPI
