+++ book/src/links.md
@@ -1,0 +1,36 @@
+<!--
+This file contains links that can be shared across pages. RustDoc links cannot currently
+be used by mdBook directly. Links are stable on crates.io so we can centralize what is
+required in this file, albeit manually.
+
+REF: https://github.com/rust-lang/mdBook/issues/1356
+REF: https://github.com/rust-lang/cargo/issues/739
+REF: https://github.com/tag1consulting/goose/issues/320
+-->
+
+[`Circuit`]: ./doc/tnc/builders/circuit_builder/struct.Circuit.html
+[`into_amplitude_network`]: ./doc/tnc/builders/circuit_builder/struct.Circuit.html#method.into_amplitude_network
+[`into_statevector_network`]: ./doc/tnc/builders/circuit_builder/struct.Circuit.html#method.into_statevector_network
+[`into_expectation_value_network`]: ./doc/tnc/builders/circuit_builder/struct.Circuit.html#method.into_expectation_value_network
+
+[`sycamore_circuit`]: ./doc/tnc/builders/sycamore_circuit/fn.sycamore_circuit.html
+
+[`import_qasm`]: ./doc/tnc/qasm/fn.import_qasm.html
+
+[`hdf5`]: ./doc/tnc/hdf5/index.html
+
+[`contraction_cost`]: ./doc/tnc/contractionpath/contraction_cost/index.html
+[`Cotengrust`]: ./doc/tnc/contractionpath/paths/cotengrust/struct.Cotengrust.html
+[`OptMethod`]: ./doc/tnc/contractionpath/paths/cotengrust/enum.OptMethod.html
+[`contract_tensor_network`]: ./doc/tnc/tensornetwork/contraction/fn.contract_tensor_network.html
+
+[`Tensor`]: ./doc/tnc/tensornetwork/tensor/struct.Tensor.html
+[`external_tensor`]: ./doc/tnc/tensornetwork/tensor/struct.Tensor.html#method.external_tensor
+[`set_tensor_data`]: ./doc/tnc/tensornetwork/tensor/struct.Tensor.html#method.set_tensor_data
+[`TensorData`]: ./doc/tnc/tensornetwork/tensordata/enum.TensorData.html
+
+[`find_partitioning`]: ./doc/tnc/tensornetwork/partitioning/fn.find_partitioning.html
+[`partition_tensor_network`]: ./doc/tnc/tensornetwork/partitioning/fn.partition_tensor_network.html
+[`IntermediatePartitioningModel`]: ./doc/tnc/contractionpath/repartitioning/simulated_annealing/struct.IntermediatePartitioningModel.html
+
+[`scatter_tensor_network`]: ./doc/tnc/mpi/communication/fn.scatter_tensor_network.html
