+++ .github/workflows/doc.yml
@@ -38,6 +38,7 @@ jobs:
           path: target/doc
   build-book:
     runs-on: ubuntu-latest
+    needs: build-doc
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # tag=v6.0.2
         with:
@@ -46,36 +47,22 @@ jobs:
         uses: peaceiris/actions-mdbook@ee69d230fe19748b7abf22df32acaa93833fad08 # tag=v2.0.0
         with:
           mdbook-version: 0.5.2
-      - name: Build book
-        run: mdbook build book
-      - name: Upload book artifact
-        uses: actions/upload-artifact@bbbca2ddaa5d8feaa63e36b76fdaad77386f024f # tag=v7.0.0
-        with:
-          name: book
-          path: book/book
-  upload-folder:
-    runs-on: ubuntu-latest
-    needs: [build-doc, build-book]
-    steps:
       - name: Download doc artifact
         uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # tag=v8.0.1
         with:
           name: doc
-          path: target/pages/doc
-      - name: Download book artifact
-        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # tag=v8.0.1
-        with:
-          name: book
-          path: target/pages/book
+          path: book/src/doc
+      - name: Build book
+        run: mdbook build book
       - name: Upload artifacts folder
         uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
         id: deployment
         with:
-          path: target/pages
+          path: book/book
   cleanup:
     runs-on: ubuntu-latest
     permissions: write-all
-    needs: upload-folder
+    needs: build-book
     steps:
       - name: Delete previous deployments
         uses: strumwolf/delete-deployment-environment@a4825dd9648c57da8437a4885c3fcad58beac69c # tag=v3.0.0
