### 2026-03-13 f0ba6df
Add script to link docs to book
