+++ book/link_docs.sh
@@ -1,0 +1,1 @@
+ln -s ../../target/doc ./src/doc
