+++ tnc/examples/repartitioning.rs
@@ -13,9 +13,7 @@ use tnc::{
         repartitioning::simulated_annealing::{self, IntermediatePartitioningModel},
     },
     tensornetwork::{
-        partitioning::{
-            find_partitioning, partition_config::PartitioningStrategy, partition_tensor_network,
-        },
+        partitioning::{find_partitioning, partition_tensor_network, PartitioningStrategy},
         tensor::Tensor,
     },
 };
