### 2026-03-13 87b9463
Use mdbook linkcheck2
Adds a mdbook backend that checks that links in the book are valid.
Since we are linking to the docs quite frequently, this helps making
sure the links are valid, especially after code refactorings.
