+++ .github/workflows/doc.yml
@@ -47,6 +47,10 @@ jobs:
         uses: peaceiris/actions-mdbook@ee69d230fe19748b7abf22df32acaa93833fad08 # tag=v2.0.0
         with:
           mdbook-version: 0.5.2
+      - name: Install mbBook-linkcheck2
+        uses: taiki-e/install-action@64c5c20c872907b6f7cd50994ac189e7274160f2 # tag=v2.68.26
+        with:
+          tool: mdbook-linkcheck2
       - name: Download doc artifact
         uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # tag=v8.0.1
         with:
