### 2026-03-13 3ddf077
Build book also on PRs
