### 2026-03-13 0d97305
Fix missing book HTML output
