+++ .github/workflows/doc.yml
@@ -63,7 +63,7 @@ jobs:
         uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
         id: deployment
         with:
-          path: book/book
+          path: book/book/html
   cleanup:
     if: ${{ github.event_name != 'pull_request' }}
     runs-on: ubuntu-latest
