+++ book/book.toml
@@ -3,5 +3,7 @@ title = "TNC - TensorNetworkContraction library"
 authors = ["Manuel Geiger"]
 language = "en"
 
+[output.html]
+
 [output.linkcheck2]
 traverse-parent-directories = true
