+++ tnc/src/io/qasm/expression_folder.rs
@@ -1,4 +1,4 @@
-use crate::qasm::{ast::Visitor, expression_simplification::fold_expr};
+use crate::io::qasm::{ast::Visitor, expression_simplification::fold_expr};
 
 /// Struct to fold constant subexpressions in an expression.
 #[derive(Debug)]
@@ -14,7 +14,7 @@ impl Visitor for ExpressionFolder {
 mod tests {
     use super::*;
 
-    use crate::qasm::ast::{BinOp, Expr, FuncType, Program, Statement, UnOp, Visitor};
+    use crate::io::qasm::ast::{BinOp, Expr, FuncType, Program, Statement, UnOp, Visitor};
 
     #[test]
     fn simplify_gatecall() {
