+++ tnc/src/lib.rs
@@ -7,8 +7,7 @@ pub mod builders;
 pub mod contractionpath;
 mod doctests;
 pub mod gates;
-pub mod hdf5;
+pub mod io;
 pub mod mpi;
-pub mod qasm;
 pub mod tensornetwork;
 mod utils;
