+++ book/src/running_on_hpc.md
@@ -25,3 +25,18 @@ Here, the steps to running this library are similar to:
 3. Build the code using `cargo build -r` (assuming the login nodes have the same architecture as the compute nodes)
 4. Write a SLURM script that calls `mpirun` on the compiled binary
 5. Submit the job
+
+## Tips
+- For max performance:
+    - Use `export RUSTFLAGS='-Ctarget-cpu=native'` when building (assuming the login nodes have the same hardware as the compute nodes)
+    - Add the following to your `Cargo.toml`:
+        ```toml
+        [profile.release]
+        codegen-units = 1 # don't split crate into multiple compilation units
+        lto = true # enable link time optimization
+        ```
+- There might be problems with finding or loading dependencies if they come from modules.
+    - You can write a custom [build script](https://doc.rust-lang.org/cargo/reference/build-scripts.html) to specify further linker args.
+- MPI might not terminate if one of the MPI ranks panics, making the job hang until it hits the time limit.
+    - Use `panic = "abort"` in the `Cargo.toml` file (see [here](https://doc.rust-lang.org/book/ch09-01-unrecoverable-errors-with-panic.html)). The library doesn't rely on unwinding panics, so aborting is fine.
+- Use `export RUST_BACKTRACE=full` to get a stacktrace in case of errors
