### 2026-04-01 ca9bafa
Bump taiki-e/install-action from 2.68.26 to 2.70.4
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.68.26 to 2.70.4.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/64c5c20c872907b6f7cd50994ac189e7274160f2...bfadeaba214680fb4ab63e710bcb2a6a17019fdc)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.70.4
  dependency-type: direct:production
  update-type: vers
[…]
