+++ .github/workflows/test.yml
@@ -172,7 +172,7 @@ jobs:
           toolchain: nightly
           components: llvm-tools-preview
       - name: cargo install cargo-llvm-cov
-        uses: taiki-e/install-action@64c5c20c872907b6f7cd50994ac189e7274160f2 # tag=v2.68.26
+        uses: taiki-e/install-action@bfadeaba214680fb4ab63e710bcb2a6a17019fdc # tag=v2.70.4
         with:
           tool: cargo-llvm-cov
       - name: cargo generate-lockfile
