+++ .github/workflows/doc.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           submodules: true
       - name: Install stable
-        uses: actions-rust-lang/setup-rust-toolchain@a0b538fa0b742a6aa35d6e2c169b4bd06d225a98 # tag=v1.15.3
+        uses: actions-rust-lang/setup-rust-toolchain@150fca883cd4034361b621bd4e6a9d34e5143606 # tag=v1.15.4
         with:
           toolchain: stable
       - name: cargo doc --no-deps
