### 2026-04-01 a9a001c
Make cotengra_check error message explicit
