+++ tnc/src/contractionpath/paths/tree_annealing.rs
@@ -31,7 +31,7 @@ impl<'a> TreeAnnealing<'a> {
         temperature_steps: Option<usize>,
         numiter: Option<usize>,
     ) -> Self {
-        assert!(cotengra_check().is_ok());
+        cotengra_check().expect("Needs python and cotengra installed");
         assert_eq!(
             minimize,
             CostType::Flops,
