+++ tnc/src/contractionpath/paths/tree_reconfiguration.rs
@@ -1,8 +1,6 @@
 use itertools::Itertools;
 use rustc_hash::FxHashMap;
-use rustengra::{
-    cotengra_check, cotengra_optimized_greedy, replace_to_ssa_path, tensor_legs_to_digit,
-};
+use rustengra::{cotengra_check, cotengra_optimized_greedy, utils::tensor_legs_to_digit};
 
 use crate::{
     contractionpath::{
@@ -65,10 +63,8 @@ impl FindPath for TreeReconfigure<'_> {
         let (inputs, outputs, size_dict) =
             tensor_legs_to_digit(&inputs, outputs.legs(), &size_dict);
 
-        let replace_path =
-            cotengra_optimized_greedy(&inputs, outputs, size_dict, self.subtree_size).unwrap();
-
-        let best_path = replace_to_ssa_path(replace_path, self.tensor.tensors().len());
+        let best_path =
+            cotengra_optimized_greedy(&inputs, &outputs, &size_dict, self.subtree_size).unwrap();
 
         self.best_path = ContractionPath::simple(best_path);
 
