### 2026-04-14 ec4ed38
Use approx_eq in tensor tests
