### 2026-04-21 #9598 @Luthaf
Hello! With help from the authors, I managed to install and test the software and did a final pass over the documentation. I did not dig too deeply in the scientific applications since I am very much not familiar with tensor networks, but the examples are working, and the Rust API looks clear enough for someone to use this library.

Overall, this is ready for publication from my side!
