### 2026-04-22 afe3663
Apply new clippy lints
