+++ tnc/src/contractionpath/paths/hyperoptimization.rs
@@ -1,6 +1,6 @@
 use itertools::Itertools;
 use rustc_hash::FxHashMap;
-use rustengra::{hyper::cotengra_hyperoptimizer, utils::tensor_legs_to_digit};
+use rustengra::hyper::cotengra_hyperoptimizer;
 
 use crate::{
     contractionpath::{
@@ -58,12 +58,9 @@ impl FindPath for Hyperoptimizer<'_> {
             },
         );
 
-        let (inputs, outputs, size_dict) =
-            tensor_legs_to_digit(&inputs, outputs.legs(), &size_dict);
-
         let best_path = cotengra_hyperoptimizer(
             &inputs,
-            &outputs,
+            outputs.legs(),
             &size_dict,
             "kahypar",
             &self.hyper_options,
