+++ tnc/src/contractionpath/paths/tree_annealing.rs
@@ -1,6 +1,6 @@
 use itertools::Itertools;
 use rustc_hash::FxHashMap;
-use rustengra::{cotengra_check, cotengra_sa_tree, utils::tensor_legs_to_digit};
+use rustengra::{cotengra_check, cotengra_sa_tree};
 
 use crate::{
     contractionpath::{
@@ -67,12 +67,9 @@ impl FindPath for TreeAnnealing<'_> {
             },
         );
 
-        let (inputs, outputs, size_dict) =
-            tensor_legs_to_digit(&inputs, outputs.legs(), &size_dict);
-
         let best_path = cotengra_sa_tree(
             &inputs,
-            &outputs,
+            outputs.legs(),
             self.temperature_steps,
             self.numiter,
             &size_dict,
