### 2026-04-22 a042f5e
Fix wrong contraction path format in TreeTempering
