### 2026-04-23 596975a
Make Hyperoptimizer work with nested tensors
