### 2026-05-01 55c98da
Bump taiki-e/install-action from 2.70.4 to 2.75.27
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.70.4 to 2.75.27.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/bfadeaba214680fb4ab63e710bcb2a6a17019fdc...7769b73c2ec98c38dfcf2e18c83cfd4880c038c1)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.75.27
  dependency-type: direct:production
  update-type: ver
[…]
