+++ .github/workflows/doc.yml
@@ -60,7 +60,7 @@ jobs:
       - name: Build book
         run: mdbook build book
       - name: Upload artifacts folder
-        uses: actions/upload-pages-artifact@7b1f4a764d45c48632c6b24a0339c27f5614fb0b # tag=v4.0.0
+        uses: actions/upload-pages-artifact@fc324d3547104276b827a68afc52ff2a11cc49c9 # tag=v5.0.0
         id: deployment
         with:
           path: book/book/html
