+++ .github/workflows/doc.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           submodules: true
       - name: Install stable
-        uses: actions-rust-lang/setup-rust-toolchain@150fca883cd4034361b621bd4e6a9d34e5143606 # tag=v1.15.4
+        uses: actions-rust-lang/setup-rust-toolchain@2b1f5e9b395427c92ee4e3331786ca3c37afe2d7 # tag=v1.16.0
         with:
           toolchain: stable
       - name: cargo doc --no-deps
