### 2026-05-08 cab5f2f
Update hyperoptimizer tests
When using the new Cotengra version, the numbers changed.
