+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2025-09-09
+## [1.0.0] - 2026-05-08
 
-Initial release
+Initial release, including
+- QASM2 importing
+- Amplitude, statevector and expectation value computation
+- Tensor network contraction
+- Tensor network partitioning
+- Rebalancing of partitioned tensor network for better time-to-solution
+- Distributed contraction of partitioned tensor networks via MPI
 
+[unreleased]: https://github.com/qc-tum/TNC/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/qc-tum/TNC/releases/tag/v1.0.0
