+++ tnc/src/gates.rs
@@ -84,7 +84,7 @@ fn matrix_transpose_inplace(data: &mut DataTensor) {
         assert!(data.ndim().is_power_of_two());
         let half = data.ndim() / 2;
         let perm = (half..data.ndim()).chain(0..half).collect_vec();
-        data.transpose(&Permutation::oneline(perm));
+        *data = data.transpose(&Permutation::oneline(perm));
     }
 }
 
@@ -160,7 +160,7 @@ impl Gate for X {
             z, o,
             o, z,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -185,7 +185,7 @@ impl Gate for Y {
             z, -i,
             i,  z,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -210,7 +210,7 @@ impl Gate for Z {
             o,  z,
             z, -o,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -234,7 +234,7 @@ impl Gate for H {
             h,  h,
             h, -h,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -260,7 +260,7 @@ impl Gate for T {
             o, z,
             z, t,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -287,7 +287,7 @@ impl Gate for U {
             (Complex64::I * phi).exp() * sin,
             (Complex64::I * (phi + lambda)).exp() * cos,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -300,7 +300,7 @@ impl Gate for U {
             -(Complex64::I * -lambda).exp() * sin,
             (Complex64::I * -(phi + lambda)).exp() * cos,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 }
 
@@ -320,7 +320,7 @@ impl Gate for Sx {
             a, b,
             b, a,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 
     fn adjoint(&self, angles: &[f64]) -> DataTensor {
@@ -347,7 +347,7 @@ impl Gate for Sy {
             a, b,
             a, a,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
+        DataTensor::new_from_flat(&[2, 2], data)
     }
 }
 
@@ -368,7 +368,7 @@ impl Gate for Sz {
             o, z,
             z, i,
         ];
-        DataTensor::new_from_flat(&[2, 2], data, None)
[…]
