+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2026-05-26
+
 ### Fixed
 - A wrong leg order bug in `CircuitBuilder` that could lead to wrong order of statevectors
 
@@ -20,5 +22,6 @@ Initial release, including
 - Rebalancing of partitioned tensor network for better time-to-solution
 - Distributed contraction of partitioned tensor networks via MPI
 
-[unreleased]: https://github.com/qc-tum/TNC/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/qc-tum/TNC/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/qc-tum/TNC/releases/tag/v1.0.1
 [1.0.0]: https://github.com/qc-tum/TNC/releases/tag/v1.0.0
