### 2026-05-26 290cd15
Use ndarray instead of tetra
Since moving from MKL to tblis made the contraction simpler, tetra
didn't comprise so much complexity anymore, becoming somewhat
superfluous. ndarray is maintained, equally efficient and can handle
arbitrary slicing, which was on the TODO list for tetra.
