+++ tnc/examples/local_contraction.rs
@@ -15,12 +15,14 @@ fn main() {
     let code = "\
 OPENQASM 2.0;
 include \"qelib1.inc\";
-
-qreg q[3];
-h q[0];
-cx q[0], q[1];
-cx q[1], q[2];
+qreg q[2];
+creg c[1];
+u2(0,0) q[0];
+u2(-pi,-pi) q[1];
+cx q[0],q[1];
+u2(-pi,-pi) q[0];
 ";
+    // let code = include_str!("/work/ga87com/MQTBench_2025-04-10-21-41-01/dj_indep_qiskit_10.qasm");
 
     // Create a Circuit instance out of the code
     let circuit = import_qasm(code);
@@ -46,5 +48,10 @@ cx q[1], q[2];
     let data = statevector.tensor_data().clone().into_data();
 
     // Print the data
-    println!("Resulting statevector is: {:?}", data.elements());
+    // println!("Resulting statevector is: {:?}", data.flatten());
+    for (i, amp) in data.flatten().iter().enumerate() {
+        if amp.norm() > 1e-6 {
+            println!("|{:02b}>: {:.4}", i, amp.norm_sqr());
+        }
+    }
 }
