+++ tnc/src/tensornetwork/contraction.rs
@@ -1,10 +1,14 @@
 //! Functionality to contract tensor networks.
 use log::debug;
-use tetra::contract;
+use ndarray::Axis;
+use tblis::{tensor_mult, TensorView};
 
 use crate::{
     contractionpath::ContractionPath,
-    tensornetwork::{tensor::Tensor, tensordata::TensorData},
+    tensornetwork::{
+        tensor::Tensor,
+        tensordata::{DataTensor, TensorData},
+    },
 };
 
 /// Fully contracts `tn` based on the given `contract_path` using ReplaceLeft format.
@@ -75,7 +79,7 @@ impl TensorContraction for Tensor {
             ..
         } = tensor_b;
 
-        let result = contract(
+        let result = contract_ndarrays(
             &tensor_symmetric_difference.legs,
             &a_legs,
             a_data.into_data(),
@@ -86,6 +90,36 @@ impl TensorContraction for Tensor {
         tensor_symmetric_difference.set_tensor_data(TensorData::Matrix(result));
         self.tensors[tensor_a_loc] = tensor_symmetric_difference;
     }
+}
+
+fn contract_ndarrays(
+    out_labels: &[usize],
+    a_labels: &[usize],
+    a_data: DataTensor,
+    b_labels: &[usize],
+    b_data: DataTensor,
+) -> DataTensor {
+    assert_eq!(a_labels.len(), a_data.ndim());
+    assert_eq!(b_labels.len(), b_data.ndim());
+
+    // Find output shape
+    let mut out_shape = Vec::with_capacity(out_labels.len());
+    for label in out_labels {
+        if let Some(a_index) = a_labels.iter().position(|l| l == label) {
+            out_shape.push(a_data.len_of(Axis(a_index)));
+        } else if let Some(b_index) = b_labels.iter().position(|l| l == label) {
+            out_shape.push(b_data.len_of(Axis(b_index)));
+        } else {
+            panic!("Out label {label} not found in input a ({a_labels:?}) or b ({b_labels:?})");
+        }
+    }
+
+    // Contract with TBLIS
+    let a_view = TensorView::new(a_labels, a_data.shape(), a_data.strides(), a_data.as_ptr());
+    let b_view = TensorView::new(b_labels, b_data.shape(), b_data.strides(), b_data.as_ptr());
+    let out_data = tensor_mult(out_labels, &out_shape, a_view, b_view);
+
+    DataTensor::from_shape_vec(out_shape, out_data).unwrap()
 }
 
 #[cfg(test)]
