+++ tnc/Cargo.toml
@@ -36,7 +36,6 @@ rayon = "1.11.0"
 ordered-float = "5.1.0"
 rustengra = { git = "https://github.com/Ectras/rustengra.git", optional = true }
 cotengrust = { git = "https://github.com/Ectras/cotengrust.git", branch = "rust_extension" }
-serde-pickle = "1.2.0"
 
 [dev-dependencies]
 mpi-test = { git = "https://github.com/Ectras/mpi-test.git" }
