+++ book/src/partitioning.md
@@ -2,7 +2,7 @@
 
 # Partitioning
 
-The library uses partitiioning of tensor networks to parallelize the contraction.
+The library uses partitioning of tensor networks to parallelize the contraction.
 For example, given the following tensor network
 
 ![A tensor network](assets/tensor_network.svg)
@@ -12,7 +12,7 @@ one can partition it into two separate tensor networks like this:
 ![A tensor network split into two networks, separated by a dashed line.](assets/partitioning.svg)
 
 To partition a tensor network, use [`find_partitioning`] and specify how many partitions should be created.
-This will use the Hypergraph partitioner KaHyPar to find roughly equally sized partitions with minimial cost for the legs between the partitions.
+This will use the Hypergraph partitioner KaHyPar to find roughly equally sized partitions with minimal cost for the legs between the partitions.
 
 To actually get a partitioned tensor network, use [`partition_tensor_network`] and pass the partitioning (you can of course also provide your own).
 The partitioning just specifies for each tensor to which partition it should belong to.
