+++ book/src/constructing_tns.md
@@ -9,7 +9,7 @@ There are multiple ways to construct tensors and tensor networks.
 Since this library focuses on simulation of quantum circuits, there are multiple methods to construct corresponding tensor networks.
 
 ### OpenQASM2 code
-If the goal is to clasically simulate quantum circuits, one can directly load OpenQASM2 code and construct a [`Circuit`] out of it using [`import_qasm`].
+If the goal is to classically simulate quantum circuits, one can directly load OpenQASM2 code and construct a [`Circuit`] out of it using [`import_qasm`].
 
 > [!NOTE]
 > This library implements many standard gates and when it encounters one in the QASM code, it will not look for a gate definition; only when it doesn't know the gate, it will decompose the gate using an earlier gate definition in the QASM code.
