### 2026-05-26 c9e4266
Add serialization test
