### 2026-05-26 5164052
Use approx crate instead of float-cmp
Only approx is supported by ndarray, so to avoid pulling two crates for
float comparison, we can just go with approx as well.
