+++ tnc/src/tensornetwork/tensordata.rs
@@ -1,7 +1,6 @@
 use std::path::PathBuf;
 
-use approx::ulps_eq;
-use float_cmp::{ApproxEq, F64Margin};
+use approx::AbsDiffEq;
 use ndarray::ArrayD;
 use num_complex::Complex64;
 use serde::{Deserialize, Serialize};
@@ -14,7 +13,7 @@ use crate::{
 pub type DataTensor = ArrayD<Complex64>;
 
 /// The data of a tensor.
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub enum TensorData {
     /// This is for composite tensors that have not been contracted yet, as well as
     /// empty tensors in general.
@@ -71,24 +70,29 @@ impl TensorData {
     }
 }
 
-impl ApproxEq for &TensorData {
-    type Margin = F64Margin;
+impl AbsDiffEq for TensorData {
+    type Epsilon = f64;
 
-    fn approx_eq<M: Into<Self::Margin>>(self, other: Self, margin: M) -> bool {
-        let margin = margin.into();
+    fn default_epsilon() -> Self::Epsilon {
+        f64::EPSILON
+    }
+
+    fn abs_diff_eq(&self, other: &Self, epsilon: Self::Epsilon) -> bool {
         match (self, other) {
             (TensorData::File(l0), TensorData::File(r0)) => l0 == r0,
             (
                 TensorData::Gate((name_l, angles_l, adjoint_l)),
                 TensorData::Gate((name_r, angles_r, adjoint_r)),
-            ) => name_l == name_r && adjoint_l == adjoint_r && angles_l.approx_eq(angles_r, margin),
+            ) => {
+                name_l == name_r
+                    && adjoint_l == adjoint_r
+                    && angles_l
+                        .iter()
+                        .zip(angles_r)
+                        .all(|(l, r)| f64::abs_diff_eq(l, r, epsilon))
+            }
             (TensorData::Matrix(l0), TensorData::Matrix(r0)) => {
-                ulps_eq!(
-                    l0,
-                    r0,
-                    max_ulps = margin.ulps as _,
-                    epsilon = margin.epsilon
-                )
+                DataTensor::abs_diff_eq(l0, r0, epsilon)
             }
             (TensorData::Uncontracted, TensorData::Uncontracted) => true,
             _ => false,
@@ -98,39 +102,39 @@ impl ApproxEq for &TensorData {
 
 #[cfg(test)]
 mod tests {
-    use float_cmp::assert_approx_eq;
+    use approx::assert_abs_diff_eq;
 
     use super::*;
 
     #[test]
-    #[should_panic(expected = "assertion failed: `(left approx_eq right)`")]
+    #[should_panic(expected = "assert_abs_diff_eq!")]
     fn gates_eq_different_name() {
         let g1 = TensorData::Gate((String::from("cx"), vec![], false));
         let g2 = TensorData::Gate((String::from("CX"), vec![], false));
-        assert_approx_eq!(&TensorData, &g1, &g2);
+        assert_abs_diff_eq!(&g1, &g2);
     }
 
     #[test]
-    #[should_panic(expected = "assertion failed: `(left approx_eq right)`")]
+    #[should_panic(expected = "assert_abs_diff_eq!")]
     fn gates_eq_adjoint() {
         let g1 = TensorData::Gate((String::from("h"), vec![], false));
         let g2 = TensorData::Gate((String::from("h"), vec![], true));
-        assert_approx_eq!(&TensorData, &g1, &g2);
[…]
