+++ tnc/Cargo.toml
@@ -21,7 +21,6 @@ itertools = "0.14.0"
 num-complex = { version = "0.4.6", features = ["serde"] }
 rand = "0.9.2"
 antlr-rust = "0.3.0-beta"
-float-cmp = "0.10.0"
 kahypar = { git = "https://github.com/Ectras/rust-kahypar" }
 permutation = "0.4.1"
 serde = { version = "1.0.218", features = ["derive"] }

[…]
