### 2026-05-26 d717bb8
Rename contract_size_tensors_exact to contract_size_tensor_bytes
