+++ tnc/src/contractionpath/repartitioning/simulated_annealing.rs
@@ -14,7 +14,7 @@ use rustc_hash::FxHashSet;
 use crate::{
     contractionpath::{
         communication_schemes::CommunicationScheme,
-        contraction_cost::{compute_memory_requirements, contract_size_tensors_exact},
+        contraction_cost::{compute_memory_requirements, contract_size_tensors_bytes},
         paths::{
             cotengrust::{Cotengrust, OptMethod},
             FindPath,
@@ -185,7 +185,7 @@ where
         let mem = compute_memory_requirements(
             partitioned_tn.tensors(),
             &path,
-            contract_size_tensors_exact,
+            contract_size_tensors_bytes,
         );
 
         if mem > limit {
