+++ CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Function `contract_size_tensors_bytes` to use as memory estimator. Computes the memory requirement for a contraction.
+
 ### Changed
 - Use TBLIS instead of HPTT + MKL for contraction. This means contractions are faster and often require only half as much memory. Also, build time decreased.
 - Use ndarray instead of own implementation for tensors. This means more features (slicing, arbitrary memory layout, ...) and interoperability.
+- Implement `approx` instead of `float-cmp` for tensors and tensor data
+
+### Removed
+- Function `contract_size_tensors_exact` (since there is no explicit transpose, the normal contraction size estimate is sufficient)
 
 ## [1.0.1] - 2026-05-26
 
