### 2026-05-27 87f810e
Add into_tensor_data method
