+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Function `contract_size_tensors_bytes` to use as memory estimator. Computes the memory requirement for a contraction.
+- Function `into_tensor_data` to get owned data from a tensor
 
 ### Changed
 - Use TBLIS instead of HPTT + MKL for contraction. This means contractions are faster and often require only half as much memory. Also, build time decreased.
