### 2026-05-28 d377b18
Add full contraction tests
