+++ tnc/tests/integration_tests.rs
@@ -1,6 +1,10 @@
+use std::f64::consts::FRAC_1_SQRT_2;
+
 use approx::assert_abs_diff_eq;
 use mpi::traits::Communicator;
 use mpi_test::mpi_test;
+use ndarray::array;
+use num_complex::Complex64;
 use rand::{rngs::StdRng, SeedableRng};
 use tnc::{
     builders::{connectivity::ConnectivityLayout, random_circuit::random_circuit},
@@ -8,6 +12,7 @@ use tnc::{
         cotengrust::{Cotengrust, OptMethod},
         FindPath,
     },
+    io::qasm::import_qasm,
     mpi::communication::{
         broadcast_path, intermediate_reduce_tensor_network, scatter_tensor_network,
     },
@@ -159,4 +164,81 @@ fn test_partitioned_contraction_need_mpi() {
 
         assert_abs_diff_eq!(&local_tn.tensor_data(), &ref_tn.tensor_data());
     }
+}
+
+#[test]
+fn dj_4qubits_statevector() {
+    let code = "OPENQASM 2.0;
+    include \"qelib1.inc\";
+    qreg q[4];
+    creg c[3];
+    u2(0,0) q[0];
+    u2(0,0) q[1];
+    h q[2];
+    u2(-pi,-pi) q[3];
+    cx q[0],q[3];
+    u2(-pi,-pi) q[0];
+    cx q[1],q[3];
+    u2(-pi,-pi) q[1];
+    cx q[2],q[3];
+    h q[2];";
+
+    let circuit = import_qasm(code);
+    let (tensor_network, permutator) = circuit.into_statevector_network();
+
+    let mut opt = Cotengrust::new(&tensor_network, OptMethod::Greedy);
+    opt.find_path();
+    let path = opt.get_best_replace_path();
+
+    let final_tensor = contract_tensor_network(tensor_network, &path);
+    let statevector = permutator.apply(final_tensor);
+    let data = statevector.into_tensor_data().into_data();
+
+    // Result is 1/sqrt(2) * (|1110> - |1111>)
+    let expected = array![
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::ZERO,
+        Complex64::new(FRAC_1_SQRT_2, 0.0),
+        Complex64::new(-FRAC_1_SQRT_2, 0.0),
+    ];
+    assert_abs_diff_eq!(data.flatten(), expected, epsilon = 1e-15);
+}
+
+#[test]
+fn qft_2qubits_expectation() {
+    let code = "OPENQASM 2.0;
+    include \"qelib1.inc\";
+    qreg q[2];
+    creg meas[2];
+    h q[1];
+    cx q[1],q[0];
+    h q[1];
+    cp(pi/2) q[1],q[0];
+    h q[0];
+    swap q[0],q[1];";
+
+    let circuit = import_qasm(code);
+    let tensor_network = circuit.into_expectation_value_network();
+
+    let mut opt = Cotengrust::new(&tensor_network, OptMethod::RandomGreedy(3));
+    opt.find_path();
+    let path = opt.get_best_replace_path();
+
+    let final_tensor = contract_tensor_network(tensor_network, &path);
+    let data = final_tensor.into_tensor_data().into_data();
+
+    let expected = array![Complex64::new(0.5, 0.0)];
+    assert_abs_diff_eq!(data.flatten(), expected, epsilon = 1e-15);
 }
