### 2026-05-28 cba3b1b
Fix typos in book
