### 2026-06-01 44cdc16
Bump taiki-e/install-action from 2.75.27 to 2.81.1
Bumps [taiki-e/install-action](https://github.com/taiki-e/install-action) from 2.75.27 to 2.81.1.
- [Release notes](https://github.com/taiki-e/install-action/releases)
- [Changelog](https://github.com/taiki-e/install-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/taiki-e/install-action/compare/7769b73c2ec98c38dfcf2e18c83cfd4880c038c1...e49978b799e49ff429d162b7a30601a569ab6538)

---
updated-dependencies:
- dependency-name: taiki-e/install-action
  dependency-version: 2.81.1
  dependency-type: direct:production
  update-type: vers
[…]
