+++ .github/workflows/doc.yml
@@ -49,7 +49,7 @@ jobs:
         with:
           mdbook-version: 0.5.2
       - name: Install mbBook-linkcheck2
-        uses: taiki-e/install-action@7769b73c2ec98c38dfcf2e18c83cfd4880c038c1 # tag=v2.75.27
+        uses: taiki-e/install-action@e49978b799e49ff429d162b7a30601a569ab6538 # tag=v2.81.1
         with:
           tool: mdbook-linkcheck2
       - name: Download doc artifact
