### 2026-06-02 a9953a2
minor change in paper.md
