### 2026-06-02 6b8697a
minor changes in paper.bib
