+++ paper/paper.bib
@@ -42,7 +42,7 @@
   title      = {Open Quantum Assembly Language},
   copyright  = {arXiv.org perpetual, non-exclusive license},
   date       = {2017},
-  doi        = {10.48550/ARXIV.1707.03429},
+  doi        = {10.48550/arxiv.1707.03429},
   eprint     = {1707.03429},
   eprinttype = {arxiv},
   publisher  = {arXiv}
@@ -67,7 +67,7 @@
   author     = {Springer, Paul and Su, Tong and Bientinesi, Paolo},
   date       = {2017},
   title      = {{HPTT}: A High-Performance Tensor Transposition {C++} Library},
-  doi        = {10.48550/ARXIV.1704.04374},
+  doi        = {10.48550/arxiv.1704.04374},
   eprint     = {1704.04374},
   eprinttype = {arxiv},
   copyright  = {arXiv.org perpetual, non-exclusive license},
@@ -82,7 +82,7 @@
   eprint        = {2507.20667},
   archiveprefix = {arXiv},
   primaryclass  = {quant-ph},
-  url           = {https://arxiv.org/abs/2507.20667}
+  doi           = {10.48550/arxiv.2507.20667}
 }
 
 @article{Arute2019,
@@ -157,7 +157,7 @@
 
 @inproceedings{QTensor,
   author    = {{Lykov}, Danylo and {Alexeev}, Yuri and {Ibrahim}, Cameron and {Galda}, Alexey},
-  title     = {{QTensor: Fast QAOA Quantum Simulator}},
+  title     = {{QTensor}: Fast QAOA Quantum Simulator},
   booktitle = {APS March Meeting Abstracts},
   year      = 2021,
   series    = {APS Meeting Abstracts},
@@ -176,7 +176,7 @@
   year          = {2024},
   archiveprefix = {arXiv},
   copyright     = {arXiv.org perpetual, non-exclusive license},
-  doi           = {10.48550/ARXIV.2408.05011},
+  doi           = {10.48550/arxiv.2408.05011},
   eprint        = {2408.05011},
   keywords      = {Quantum Physics (quant-ph), FOS: Physical sciences},
   primaryclass  = {quant-ph},
