### 2026-06-02 60da0f0
Add JOSS publication information
