### 2026-06-03 3d3486f
Return result struct from find_path
