+++ tnc/src/contractionpath/paths/weighted_branchbound.rs
@@ -7,8 +7,8 @@ use crate::{
     contractionpath::{
         candidates::Candidate,
         contraction_cost::{contract_op_cost_tensors, contract_size_tensors},
-        paths::{CostType, FindPath},
-        ssa_ordering, ssa_replace_ordering, ContractionPath,
+        paths::{BasicContractionPathResult, ContractionPathResult, CostType, Pathfinder},
+        ssa_ordering, ContractionPath,
     },
     tensornetwork::tensor::Tensor,
     utils::traits::HashMapInsertNew,
@@ -168,10 +168,12 @@ impl<'a> WeightedBranchBound<'a> {
     }
 }
 
-impl FindPath for WeightedBranchBound<'_> {
-    fn find_path(&mut self) {
+impl Pathfinder for WeightedBranchBound<'_> {
+    type Result = BasicContractionPathResult;
+
+    fn find_path(&mut self) -> BasicContractionPathResult {
         if self.tn.is_leaf() {
-            return;
+            return BasicContractionPathResult::default();
         }
         let tensors = self.tn.tensors().clone();
         self.result_cache.clear();
@@ -194,34 +196,23 @@ impl FindPath for WeightedBranchBound<'_> {
                     self.comm_cache.clone(),
                     self.minimize,
                 );
-                bb.find_path();
-                nested_paths.insert(index, bb.get_best_path().clone());
+                let result = bb.find_path();
+                nested_paths.insert(index, result.ssa_path().clone());
                 tensor = tensor.external_tensor();
             }
             self.tensor_cache.insert_new(index, tensor);
         }
         let remaining = (0..self.tn.tensors().len()).collect_vec();
         self.branch_iterate(&[], &remaining, 0f64, 0f64);
-        self.best_path = ContractionPath {
+        let best_path = ContractionPath {
             nested: nested_paths,
             toplevel: std::mem::take(&mut self.best_path).into_simple(),
         };
-    }
-
-    fn get_best_flops(&self) -> f64 {
-        self.best_flops
-    }
-
-    fn get_best_size(&self) -> f64 {
-        self.best_size
-    }
-
-    fn get_best_path(&self) -> &ContractionPath {
-        &self.best_path
-    }
-
-    fn get_best_replace_path(&self) -> ContractionPath {
-        ssa_replace_ordering(&self.best_path)
+        BasicContractionPathResult {
+            ssa_path: best_path,
+            flops: self.best_flops,
+            size: self.best_size,
+        }
     }
 }
 
@@ -232,7 +223,7 @@ mod tests {
     use rustc_hash::FxHashMap;
 
     use crate::contractionpath::paths::CostType;
-    use crate::contractionpath::paths::FindPath;
+    use crate::contractionpath::paths::Pathfinder;
     use crate::path;
     use crate::tensornetwork::tensor::Tensor;
 
@@ -281,26 +272,31 @@ mod tests {
     fn test_contract_order_simple() {
         let (tn, latency_costs) = setup_simple();
         let mut opt = WeightedBranchBound::new(&tn, None, 20., latency_costs, CostType::Flops);
-        opt.find_path();
+        let result = opt.find_path();
 
-        assert_eq!(opt.best_flops, 640.);
-        assert_eq!(opt.best_size, 538.);
-        assert_eq!(opt.get_best_path(), &path![(1, 0), (2, 3)]);
-        assert_eq!(opt.get_best_replace_path(), path![(1, 0), (2, 1)]);
+        assert_eq!(
+            result,
+            BasicContractionPathResult {
+                ssa_path: path![(1, 0), (2, 3)],
+                flops: 640.,
+                size: 538.
+            }
[…]
