+++ tnc/src/contractionpath/paths/hyperoptimization.rs
@@ -5,7 +5,7 @@ use rustengra::hyper::cotengra_hyperoptimizer;
 use crate::{
     contractionpath::{
         contraction_cost::contract_path_cost,
-        paths::{CostType, FindPath},
+        paths::{BasicContractionPathResult, ContractionPathResult, CostType, Pathfinder},
         ssa_replace_ordering, ContractionPath,
     },
     tensornetwork::tensor::Tensor,
@@ -18,9 +18,6 @@ pub use rustengra::hyper::HyperOptions;
 pub struct Hyperoptimizer<'a> {
     tensor: &'a Tensor,
     hyper_options: HyperOptions,
-    best_flops: f64,
-    best_size: f64,
-    best_path: ContractionPath,
 }
 
 impl<'a> Hyperoptimizer<'a> {
@@ -33,15 +30,14 @@ impl<'a> Hyperoptimizer<'a> {
         Self {
             tensor,
             hyper_options,
-            best_flops: f64::INFINITY,
-            best_size: f64::INFINITY,
-            best_path: ContractionPath::default(),
         }
     }
 }
 
-impl FindPath for Hyperoptimizer<'_> {
-    fn find_path(&mut self) {
+impl Pathfinder for Hyperoptimizer<'_> {
+    type Result = BasicContractionPathResult;
+
+    fn find_path(&mut self) -> BasicContractionPathResult {
         // Handle nested tensors first
         let mut nested_paths = FxHashMap::default();
         let inputs = self
@@ -53,8 +49,8 @@ impl FindPath for Hyperoptimizer<'_> {
                 if tensor.is_composite() {
                     let mut hp =
                         Hyperoptimizer::new(tensor, CostType::Flops, self.hyper_options.clone());
-                    hp.find_path();
-                    nested_paths.insert(index, hp.get_best_path().clone());
+                    let result = hp.find_path();
+                    nested_paths.insert(index, result.ssa_path().clone());
                     tensor.external_tensor().legs
                 } else {
                     tensor.legs.clone()
@@ -80,32 +76,19 @@ impl FindPath for Hyperoptimizer<'_> {
         )
         .unwrap();
 
-        self.best_path = ContractionPath {
+        let best_path = ContractionPath {
             nested: nested_paths,
             toplevel: ssa_path,
         };
+        let replace_path = ssa_replace_ordering(&best_path);
 
-        let (op_cost, mem_cost) =
-            contract_path_cost(self.tensor.tensors(), &self.get_best_replace_path(), true);
+        let (op_cost, mem_cost) = contract_path_cost(self.tensor.tensors(), &&replace_path, true);
 
-        self.best_flops = op_cost;
-        self.best_size = mem_cost;
-    }
-
-    fn get_best_flops(&self) -> f64 {
-        self.best_flops
-    }
-
-    fn get_best_size(&self) -> f64 {
-        self.best_size
-    }
-
-    fn get_best_path(&self) -> &ContractionPath {
-        &self.best_path
-    }
-
-    fn get_best_replace_path(&self) -> ContractionPath {
-        ssa_replace_ordering(&self.best_path)
+        BasicContractionPathResult {
+            ssa_path: best_path,
+            flops: op_cost,
+            size: mem_cost,
+        }
     }
 }
 
@@ -116,7 +99,7 @@ mod tests {
     use rustc_hash::FxHashMap;
 
     use crate::{
-        contractionpath::paths::{CostType, FindPath},
+        contractionpath::paths::{CostType, Pathfinder},
         path,
         tensornetwork::tensor::Tensor,
     };
@@ -166,12 +149,16 @@ mod tests {
                 .with_max_repeats(16)
[…]
