+++ tnc/src/contractionpath/repartitioning.rs
@@ -10,7 +10,7 @@ use crate::{
         contraction_cost::{communication_path_op_costs, contract_path_cost},
         paths::{
             cotengrust::{Cotengrust, OptMethod},
-            FindPath,
+            ContractionPathResult, Pathfinder,
         },
         ContractionPath,
     },
@@ -36,8 +36,8 @@ where
 
     // Find contraction path
     let mut greedy = Cotengrust::new(&partitioned_tn, OptMethod::Greedy);
-    greedy.find_path();
-    let path = greedy.get_best_replace_path();
+    let result = greedy.find_path();
+    let path = result.replace_path();
 
     // Store the local paths (and costs)
     let mut latency_map =
