### 2026-06-03 9b943be
Add tensor argument to find_path
