### 2026-06-03 a9c518b
Reuse path finder for nested tensor networks
