+++ benchmark/Cargo.toml
@@ -8,7 +8,7 @@ rand = "0.9.2"
 clap = { version = "4.5.28", features = ["derive"] }
 flate2 = "1.1.1"
 itertools = "0.14.0"
-tnc = { version = "1.0.0", path = "../tnc", features = ["cotengra"] }
+tnc = { path = "../tnc", features = ["cotengra"] }
 bincode = { version = "2.0.1", features = ["serde"] }
 mpi = "0.8.1"
 serde_json = "1.0.133"
