+++ tnc/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tnc"
-version = "1.0.0"
+version = "1.0.1"
 edition = "2021"
 description = "A distributed-memory tensor network contraction library using partitioning."
 license = "MIT OR Apache-2.0"
