### 2026-06-03 9878c1e
Apply clippy lints
